Community matrix and phylogeny of Acacia in Australia

Dataset of Australian Acacia from Mishler et al. 2014 (Nat. Comm.)

acacia

Format

A list with two elements:

phy

Phylogeny of Australian Acacia (list of class "DNAbin"). Tip labels are specific epithet, except for the outgroup, which includes genus and specific epithet. Includes 508 ingroup taxa (genus Acacia) and two outgroup taxa.

comm

Community matrix of Australian Acacia (dataframe). Column names are specific epithet of each species. Row names are centroids of 50km grid cells in Australian Albers equal area EPSG:3577 projection. 3037 rows (sites) x 506 columns (species). Data are counts, i.e., the number of times a species was observed in a grid cell.

Source

Mishler, B., Knerr, N., González-Orozco, C. et al. Phylogenetic measures of biodiversity and neo- and paleo-endemism in Australian Acacia. Nat Commun 5, 4473 (2014). https://doi.org/10.1038/ncomms5473

Examples

# Example phylogeny
acacia$phy

#> $edge
#>         [,1] [,2]
#>    [1,]  511  512
#>    [2,]  512    1
#>    [3,]  512    2
#>    [4,]  511  513
#>    [5,]  513  514
#>    [6,]  514  515
#>    [7,]  515  516
#>    [8,]  516    3
#>    [9,]  516    4
#>   [10,]  515  517
#>   [11,]  517    5
#>   [12,]  517  518
#>   [13,]  518    6
#>   [14,]  518    7
#>   [15,]  514  519
#>   [16,]  519  520
#>   [17,]  520  521
#>   [18,]  521    8
#>   [19,]  521    9
#>   [20,]  520  522
#>   [21,]  522   10
#>   [22,]  522  523
#>   [23,]  523   11
#>   [24,]  523   12
#>   [25,]  519  524
#>   [26,]  524   13
#>   [27,]  524  525
#>   [28,]  525   14
#>   [29,]  525  526
#>   [30,]  526   15
#>   [31,]  526  527
#>   [32,]  527   16
#>   [33,]  527  528
#>   [34,]  528   17
#>   [35,]  528   18
#>   [36,]  513  529
#>   [37,]  529  530
#>   [38,]  530   19
#>   [39,]  530  531
#>   [40,]  531   20
#>   [41,]  531   21
#>   [42,]  529  532
#>   [43,]  532  533
#>   [44,]  533  534
#>   [45,]  534   22
#>   [46,]  534  535
#>   [47,]  535   23
#>   [48,]  535   24
#>   [49,]  533  536
#>   [50,]  536  537
#>   [51,]  537   25
#>   [52,]  537  538
#>   [53,]  538   26
#>   [54,]  538  539
#>   [55,]  539   27
#>   [56,]  539   28
#>   [57,]  536  540
#>   [58,]  540   29
#>   [59,]  540  541
#>   [60,]  541   30
#>   [61,]  541  542
#>   [62,]  542   31
#>   [63,]  542  543
#>   [64,]  543   32
#>   [65,]  543   33
#>   [66,]  532  544
#>   [67,]  544  545
#>   [68,]  545  546
#>   [69,]  546  547
#>   [70,]  547  548
#>   [71,]  548  549
#>   [72,]  549   34
#>   [73,]  549   35
#>   [74,]  548  550
#>   [75,]  550   36
#>   [76,]  550   37
#>   [77,]  547  551
#>   [78,]  551  552
#>   [79,]  552   38
#>   [80,]  552   39
#>   [81,]  551  553
#>   [82,]  553   40
#>   [83,]  553   41
#>   [84,]  546  554
#>   [85,]  554  555
#>   [86,]  555   42
#>   [87,]  555  556
#>   [88,]  556   43
#>   [89,]  556   44
#>   [90,]  554  557
#>   [91,]  557   45
#>   [92,]  557  558
#>   [93,]  558  559
#>   [94,]  559   46
#>   [95,]  559   47
#>   [96,]  558  560
#>   [97,]  560  561
#>   [98,]  561   48
#>   [99,]  561   49
#>  [100,]  560  562
#>  [101,]  562  563
#>  [102,]  563   50
#>  [103,]  563   51
#>  [104,]  562  564
#>  [105,]  564   52
#>  [106,]  564  565
#>  [107,]  565   53
#>  [108,]  565  566
#>  [109,]  566   54
#>  [110,]  566  567
#>  [111,]  567   55
#>  [112,]  567   56
#>  [113,]  545  568
#>  [114,]  568  569
#>  [115,]  569  570
#>  [116,]  570   57
#>  [117,]  570  571
#>  [118,]  571   58
#>  [119,]  571   59
#>  [120,]  569  572
#>  [121,]  572   60
#>  [122,]  572  573
#>  [123,]  573   61
#>  [124,]  573  574
#>  [125,]  574  575
#>  [126,]  575   62
#>  [127,]  575   63
#>  [128,]  574  576
#>  [129,]  576   64
#>  [130,]  576  577
#>  [131,]  577   65
#>  [132,]  577   66
#>  [133,]  568  578
#>  [134,]  578  579
#>  [135,]  579   67
#>  [136,]  579   68
#>  [137,]  578  580
#>  [138,]  580  581
#>  [139,]  581   69
#>  [140,]  581   70
#>  [141,]  580  582
#>  [142,]  582  583
#>  [143,]  583   71
#>  [144,]  583  584
#>  [145,]  584   72
#>  [146,]  584  585
#>  [147,]  585   73
#>  [148,]  585  586
#>  [149,]  586   74
#>  [150,]  586  587
#>  [151,]  587   75
#>  [152,]  587  588
#>  [153,]  588   76
#>  [154,]  588   77
#>  [155,]  582  589
#>  [156,]  589  590
#>  [157,]  590  591
#>  [158,]  591   78
#>  [159,]  591   79
#>  [160,]  590  592
#>  [161,]  592   80
#>  [162,]  592  593
#>  [163,]  593   81
#>  [164,]  593   82
#>  [165,]  589  594
#>  [166,]  594   83
#>  [167,]  594  595
#>  [168,]  595   84
#>  [169,]  595  596
#>  [170,]  596  597
#>  [171,]  597   85
#>  [172,]  597  598
#>  [173,]  598   86
#>  [174,]  598   87
#>  [175,]  596  599
#>  [176,]  599  600
#>  [177,]  600   88
#>  [178,]  600   89
#>  [179,]  599  601
#>  [180,]  601   90
#>  [181,]  601   91
#>  [182,]  544  602
#>  [183,]  602  603
#>  [184,]  603  604
#>  [185,]  604  605
#>  [186,]  605   92
#>  [187,]  605   93
#>  [188,]  604  606
#>  [189,]  606   94
#>  [190,]  606  607
#>  [191,]  607   95
#>  [192,]  607   96
#>  [193,]  603  608
#>  [194,]  608  609
#>  [195,]  609   97
#>  [196,]  609  610
#>  [197,]  610   98
#>  [198,]  610   99
#>  [199,]  608  611
#>  [200,]  611  612
#>  [201,]  612  100
#>  [202,]  612  101
#>  [203,]  611  613
#>  [204,]  613  102
#>  [205,]  613  614
#>  [206,]  614  103
#>  [207,]  614  104
#>  [208,]  602  615
#>  [209,]  615  616
#>  [210,]  616  105
#>  [211,]  616  106
#>  [212,]  615  617
#>  [213,]  617  618
#>  [214,]  618  619
#>  [215,]  619  107
#>  [216,]  619  620
#>  [217,]  620  108
#>  [218,]  620  621
#>  [219,]  621  109
#>  [220,]  621  622
#>  [221,]  622  110
#>  [222,]  622  623
#>  [223,]  623  111
#>  [224,]  623  112
#>  [225,]  618  624
#>  [226,]  624  625
#>  [227,]  625  626
#>  [228,]  626  113
#>  [229,]  626  114
#>  [230,]  625  627
#>  [231,]  627  628
#>  [232,]  628  115
#>  [233,]  628  116
#>  [234,]  627  629
#>  [235,]  629  117
#>  [236,]  629  630
#>  [237,]  630  118
#>  [238,]  630  119
#>  [239,]  624  631
#>  [240,]  631  632
#>  [241,]  632  120
#>  [242,]  632  121
#>  [243,]  631  633
#>  [244,]  633  634
#>  [245,]  634  122
#>  [246,]  634  123
#>  [247,]  633  635
#>  [248,]  635  124
#>  [249,]  635  636
#>  [250,]  636  125
#>  [251,]  636  637
#>  [252,]  637  126
#>  [253,]  637  638
#>  [254,]  638  639
#>  [255,]  639  127
#>  [256,]  639  128
#>  [257,]  638  640
#>  [258,]  640  129
#>  [259,]  640  641
#>  [260,]  641  130
#>  [261,]  641  131
#>  [262,]  617  642
#>  [263,]  642  132
#>  [264,]  642  643
#>  [265,]  643  644
#>  [266,]  644  645
#>  [267,]  645  133
#>  [268,]  645  646
#>  [269,]  646  647
#>  [270,]  647  134
#>  [271,]  647  648
#>  [272,]  648  649
#>  [273,]  649  135
#>  [274,]  649  650
#>  [275,]  650  136
#>  [276,]  650  137
#>  [277,]  648  651
#>  [278,]  651  138
#>  [279,]  651  652
#>  [280,]  652  139
#>  [281,]  652  653
#>  [282,]  653  140
#>  [283,]  653  141
#>  [284,]  646  654
#>  [285,]  654  142
#>  [286,]  654  655
#>  [287,]  655  656
#>  [288,]  656  143
#>  [289,]  656  657
#>  [290,]  657  658
#>  [291,]  658  144
#>  [292,]  658  145
#>  [293,]  657  659
#>  [294,]  659  146
#>  [295,]  659  147
#>  [296,]  655  660
#>  [297,]  660  148
#>  [298,]  660  661
#>  [299,]  661  149
#>  [300,]  661  662
#>  [301,]  662  663
#>  [302,]  663  150
#>  [303,]  663  664
#>  [304,]  664  151
#>  [305,]  664  665
#>  [306,]  665  152
#>  [307,]  665  153
#>  [308,]  662  666
#>  [309,]  666  154
#>  [310,]  666  667
#>  [311,]  667  155
#>  [312,]  667  668
#>  [313,]  668  156
#>  [314,]  668  157
#>  [315,]  644  669
#>  [316,]  669  670
#>  [317,]  670  158
#>  [318,]  670  671
#>  [319,]  671  159
#>  [320,]  671  160
#>  [321,]  669  672
#>  [322,]  672  673
#>  [323,]  673  161
#>  [324,]  673  162
#>  [325,]  672  674
#>  [326,]  674  675
#>  [327,]  675  676
#>  [328,]  676  163
#>  [329,]  676  677
#>  [330,]  677  164
#>  [331,]  677  165
#>  [332,]  675  678
#>  [333,]  678  679
#>  [334,]  679  166
#>  [335,]  679  680
#>  [336,]  680  167
#>  [337,]  680  681
#>  [338,]  681  168
#>  [339,]  681  169
#>  [340,]  678  682
#>  [341,]  682  683
#>  [342,]  683  170
#>  [343,]  683  684
#>  [344,]  684  171
#>  [345,]  684  172
#>  [346,]  682  685
#>  [347,]  685  686
#>  [348,]  686  687
#>  [349,]  687  173
#>  [350,]  687  174
#>  [351,]  686  688
#>  [352,]  688  175
#>  [353,]  688  689
#>  [354,]  689  176
#>  [355,]  689  690
#>  [356,]  690  177
#>  [357,]  690  178
#>  [358,]  685  691
#>  [359,]  691  179
#>  [360,]  691  692
#>  [361,]  692  693
#>  [362,]  693  180
#>  [363,]  693  181
#>  [364,]  692  694
#>  [365,]  694  695
#>  [366,]  695  182
#>  [367,]  695  696
#>  [368,]  696  183
#>  [369,]  696  697
#>  [370,]  697  184
#>  [371,]  697  185
#>  [372,]  694  698
#>  [373,]  698  699
#>  [374,]  699  186
#>  [375,]  699  187
#>  [376,]  698  700
#>  [377,]  700  701
#>  [378,]  701  188
#>  [379,]  701  189
#>  [380,]  700  702
#>  [381,]  702  190
#>  [382,]  702  703
#>  [383,]  703  191
#>  [384,]  703  192
#>  [385,]  674  704
#>  [386,]  704  705
#>  [387,]  705  193
#>  [388,]  705  706
#>  [389,]  706  194
#>  [390,]  706  195
#>  [391,]  704  707
#>  [392,]  707  196
#>  [393,]  707  708
#>  [394,]  708  197
#>  [395,]  708  709
#>  [396,]  709  710
#>  [397,]  710  198
#>  [398,]  710  711
#>  [399,]  711  712
#>  [400,]  712  199
#>  [401,]  712  200
#>  [402,]  711  713
#>  [403,]  713  201
#>  [404,]  713  202
#>  [405,]  709  714
#>  [406,]  714  715
#>  [407,]  715  203
#>  [408,]  715  204
#>  [409,]  714  716
#>  [410,]  716  717
#>  [411,]  717  205
#>  [412,]  717  206
#>  [413,]  716  718
#>  [414,]  718  719
#>  [415,]  719  720
#>  [416,]  720  721
#>  [417,]  721  207
#>  [418,]  721  208
#>  [419,]  720  722
#>  [420,]  722  209
#>  [421,]  722  723
#>  [422,]  723  210
#>  [423,]  723  211
#>  [424,]  719  724
#>  [425,]  724  725
#>  [426,]  725  212
#>  [427,]  725  726
#>  [428,]  726  213
#>  [429,]  726  214
#>  [430,]  724  727
#>  [431,]  727  728
#>  [432,]  728  215
#>  [433,]  728  216
#>  [434,]  727  729
#>  [435,]  729  217
#>  [436,]  729  730
#>  [437,]  730  218
#>  [438,]  730  219
#>  [439,]  718  731
#>  [440,]  731  732
#>  [441,]  732  220
#>  [442,]  732  733
#>  [443,]  733  734
#>  [444,]  734  221
#>  [445,]  734  222
#>  [446,]  733  735
#>  [447,]  735  736
#>  [448,]  736  223
#>  [449,]  736  737
#>  [450,]  737  224
#>  [451,]  737  738
#>  [452,]  738  225
#>  [453,]  738  739
#>  [454,]  739  226
#>  [455,]  739  740
#>  [456,]  740  227
#>  [457,]  740  741
#>  [458,]  741  228
#>  [459,]  741  229
#>  [460,]  735  742
#>  [461,]  742  230
#>  [462,]  742  743
#>  [463,]  743  744
#>  [464,]  744  231
#>  [465,]  744  745
#>  [466,]  745  232
#>  [467,]  745  233
#>  [468,]  743  746
#>  [469,]  746  234
#>  [470,]  746  747
#>  [471,]  747  235
#>  [472,]  747  748
#>  [473,]  748  236
#>  [474,]  748  749
#>  [475,]  749  237
#>  [476,]  749  750
#>  [477,]  750  238
#>  [478,]  750  751
#>  [479,]  751  239
#>  [480,]  751  752
#>  [481,]  752  240
#>  [482,]  752  241
#>  [483,]  731  753
#>  [484,]  753  754
#>  [485,]  754  242
#>  [486,]  754  755
#>  [487,]  755  243
#>  [488,]  755  244
#>  [489,]  753  756
#>  [490,]  756  245
#>  [491,]  756  757
#>  [492,]  757  246
#>  [493,]  757  758
#>  [494,]  758  759
#>  [495,]  759  247
#>  [496,]  759  760
#>  [497,]  760  761
#>  [498,]  761  248
#>  [499,]  761  762
#>  [500,]  762  249
#>  [501,]  762  250
#>  [502,]  760  763
#>  [503,]  763  251
#>  [504,]  763  764
#>  [505,]  764  252
#>  [506,]  764  765
#>  [507,]  765  253
#>  [508,]  765  766
#>  [509,]  766  254
#>  [510,]  766  255
#>  [511,]  758  767
#>  [512,]  767  768
#>  [513,]  768  256
#>  [514,]  768  257
#>  [515,]  767  769
#>  [516,]  769  258
#>  [517,]  769  770
#>  [518,]  770  771
#>  [519,]  771  259
#>  [520,]  771  772
#>  [521,]  772  260
#>  [522,]  772  773
#>  [523,]  773  774
#>  [524,]  774  261
#>  [525,]  774  262
#>  [526,]  773  775
#>  [527,]  775  263
#>  [528,]  775  264
#>  [529,]  770  776
#>  [530,]  776  777
#>  [531,]  777  265
#>  [532,]  777  778
#>  [533,]  778  266
#>  [534,]  778  779
#>  [535,]  779  267
#>  [536,]  779  268
#>  [537,]  776  780
#>  [538,]  780  781
#>  [539,]  781  269
#>  [540,]  781  270
#>  [541,]  780  782
#>  [542,]  782  783
#>  [543,]  783  271
#>  [544,]  783  784
#>  [545,]  784  272
#>  [546,]  784  273
#>  [547,]  782  785
#>  [548,]  785  274
#>  [549,]  785  786
#>  [550,]  786  275
#>  [551,]  786  787
#>  [552,]  787  276
#>  [553,]  787  277
#>  [554,]  643  788
#>  [555,]  788  789
#>  [556,]  789  790
#>  [557,]  790  278
#>  [558,]  790  279
#>  [559,]  789  791
#>  [560,]  791  280
#>  [561,]  791  792
#>  [562,]  792  281
#>  [563,]  792  282
#>  [564,]  788  793
#>  [565,]  793  794
#>  [566,]  794  795
#>  [567,]  795  796
#>  [568,]  796  283
#>  [569,]  796  284
#>  [570,]  795  797
#>  [571,]  797  798
#>  [572,]  798  285
#>  [573,]  798  286
#>  [574,]  797  799
#>  [575,]  799  287
#>  [576,]  799  800
#>  [577,]  800  288
#>  [578,]  800  801
#>  [579,]  801  289
#>  [580,]  801  802
#>  [581,]  802  290
#>  [582,]  802  803
#>  [583,]  803  291
#>  [584,]  803  804
#>  [585,]  804  292
#>  [586,]  804  293
#>  [587,]  794  805
#>  [588,]  805  806
#>  [589,]  806  807
#>  [590,]  807  294
#>  [591,]  807  295
#>  [592,]  806  808
#>  [593,]  808  296
#>  [594,]  808  809
#>  [595,]  809  297
#>  [596,]  809  810
#>  [597,]  810  298
#>  [598,]  810  811
#>  [599,]  811  299
#>  [600,]  811  300
#>  [601,]  805  812
#>  [602,]  812  813
#>  [603,]  813  814
#>  [604,]  814  815
#>  [605,]  815  301
#>  [606,]  815  302
#>  [607,]  814  816
#>  [608,]  816  303
#>  [609,]  816  304
#>  [610,]  813  817
#>  [611,]  817  305
#>  [612,]  817  818
#>  [613,]  818  819
#>  [614,]  819  306
#>  [615,]  819  307
#>  [616,]  818  820
#>  [617,]  820  308
#>  [618,]  820  821
#>  [619,]  821  309
#>  [620,]  821  310
#>  [621,]  812  822
#>  [622,]  822  823
#>  [623,]  823  311
#>  [624,]  823  312
#>  [625,]  822  824
#>  [626,]  824  825
#>  [627,]  825  313
#>  [628,]  825  826
#>  [629,]  826  827
#>  [630,]  827  314
#>  [631,]  827  828
#>  [632,]  828  315
#>  [633,]  828  316
#>  [634,]  826  829
#>  [635,]  829  317
#>  [636,]  829  830
#>  [637,]  830  318
#>  [638,]  830  319
#>  [639,]  824  831
#>  [640,]  831  832
#>  [641,]  832  833
#>  [642,]  833  320
#>  [643,]  833  321
#>  [644,]  832  834
#>  [645,]  834  322
#>  [646,]  834  835
#>  [647,]  835  323
#>  [648,]  835  836
#>  [649,]  836  837
#>  [650,]  837  324
#>  [651,]  837  838
#>  [652,]  838  325
#>  [653,]  838  326
#>  [654,]  836  839
#>  [655,]  839  327
#>  [656,]  839  840
#>  [657,]  840  841
#>  [658,]  841  842
#>  [659,]  842  328
#>  [660,]  842  329
#>  [661,]  841  843
#>  [662,]  843  330
#>  [663,]  843  331
#>  [664,]  840  844
#>  [665,]  844  845
#>  [666,]  845  332
#>  [667,]  845  333
#>  [668,]  844  846
#>  [669,]  846  847
#>  [670,]  847  334
#>  [671,]  847  335
#>  [672,]  846  848
#>  [673,]  848  849
#>  [674,]  849  336
#>  [675,]  849  850
#>  [676,]  850  337
#>  [677,]  850  338
#>  [678,]  848  851
#>  [679,]  851  339
#>  [680,]  851  852
#>  [681,]  852  340
#>  [682,]  852  853
#>  [683,]  853  341
#>  [684,]  853  342
#>  [685,]  831  854
#>  [686,]  854  343
#>  [687,]  854  855
#>  [688,]  855  856
#>  [689,]  856  344
#>  [690,]  856  857
#>  [691,]  857  345
#>  [692,]  857  346
#>  [693,]  855  858
#>  [694,]  858  859
#>  [695,]  859  860
#>  [696,]  860  347
#>  [697,]  860  861
#>  [698,]  861  348
#>  [699,]  861  349
#>  [700,]  859  862
#>  [701,]  862  350
#>  [702,]  862  863
#>  [703,]  863  351
#>  [704,]  863  864
#>  [705,]  864  352
#>  [706,]  864  353
#>  [707,]  858  865
#>  [708,]  865  354
#>  [709,]  865  866
#>  [710,]  866  867
#>  [711,]  867  355
#>  [712,]  867  868
#>  [713,]  868  356
#>  [714,]  868  869
#>  [715,]  869  357
#>  [716,]  869  870
#>  [717,]  870  358
#>  [718,]  870  359
#>  [719,]  866  871
#>  [720,]  871  872
#>  [721,]  872  360
#>  [722,]  872  873
#>  [723,]  873  361
#>  [724,]  873  362
#>  [725,]  871  874
#>  [726,]  874  875
#>  [727,]  875  363
#>  [728,]  875  364
#>  [729,]  874  876
#>  [730,]  876  365
#>  [731,]  876  877
#>  [732,]  877  366
#>  [733,]  877  367
#>  [734,]  793  878
#>  [735,]  878  879
#>  [736,]  879  880
#>  [737,]  880  368
#>  [738,]  880  369
#>  [739,]  879  881
#>  [740,]  881  370
#>  [741,]  881  882
#>  [742,]  882  371
#>  [743,]  882  372
#>  [744,]  878  883
#>  [745,]  883  373
#>  [746,]  883  884
#>  [747,]  884  885
#>  [748,]  885  374
#>  [749,]  885  375
#>  [750,]  884  886
#>  [751,]  886  376
#>  [752,]  886  887
#>  [753,]  887  888
#>  [754,]  888  889
#>  [755,]  889  377
#>  [756,]  889  890
#>  [757,]  890  378
#>  [758,]  890  891
#>  [759,]  891  379
#>  [760,]  891  892
#>  [761,]  892  380
#>  [762,]  892  893
#>  [763,]  893  381
#>  [764,]  893  382
#>  [765,]  888  894
#>  [766,]  894  895
#>  [767,]  895  383
#>  [768,]  895  896
#>  [769,]  896  897
#>  [770,]  897  384
#>  [771,]  897  385
#>  [772,]  896  898
#>  [773,]  898  386
#>  [774,]  898  899
#>  [775,]  899  387
#>  [776,]  899  388
#>  [777,]  894  900
#>  [778,]  900  901
#>  [779,]  901  902
#>  [780,]  902  389
#>  [781,]  902  903
#>  [782,]  903  390
#>  [783,]  903  904
#>  [784,]  904  391
#>  [785,]  904  392
#>  [786,]  901  905
#>  [787,]  905  906
#>  [788,]  906  393
#>  [789,]  906  394
#>  [790,]  905  907
#>  [791,]  907  395
#>  [792,]  907  908
#>  [793,]  908  396
#>  [794,]  908  909
#>  [795,]  909  397
#>  [796,]  909  910
#>  [797,]  910  398
#>  [798,]  910  399
#>  [799,]  900  911
#>  [800,]  911  912
#>  [801,]  912  400
#>  [802,]  912  913
#>  [803,]  913  914
#>  [804,]  914  401
#>  [805,]  914  402
#>  [806,]  913  915
#>  [807,]  915  403
#>  [808,]  915  404
#>  [809,]  911  916
#>  [810,]  916  917
#>  [811,]  917  405
#>  [812,]  917  918
#>  [813,]  918  406
#>  [814,]  918  407
#>  [815,]  916  919
#>  [816,]  919  408
#>  [817,]  919  920
#>  [818,]  920  409
#>  [819,]  920  921
#>  [820,]  921  410
#>  [821,]  921  922
#>  [822,]  922  411
#>  [823,]  922  923
#>  [824,]  923  412
#>  [825,]  923  413
#>  [826,]  887  924
#>  [827,]  924  925
#>  [828,]  925  414
#>  [829,]  925  926
#>  [830,]  926  415
#>  [831,]  926  927
#>  [832,]  927  416
#>  [833,]  927  928
#>  [834,]  928  417
#>  [835,]  928  929
#>  [836,]  929  418
#>  [837,]  929  930
#>  [838,]  930  419
#>  [839,]  930  420
#>  [840,]  924  931
#>  [841,]  931  421
#>  [842,]  931  932
#>  [843,]  932  933
#>  [844,]  933  934
#>  [845,]  934  422
#>  [846,]  934  935
#>  [847,]  935  936
#>  [848,]  936  423
#>  [849,]  936  937
#>  [850,]  937  424
#>  [851,]  937  938
#>  [852,]  938  425
#>  [853,]  938  426
#>  [854,]  935  939
#>  [855,]  939  940
#>  [856,]  940  941
#>  [857,]  941  427
#>  [858,]  941  428
#>  [859,]  940  942
#>  [860,]  942  943
#>  [861,]  943  429
#>  [862,]  943  430
#>  [863,]  942  944
#>  [864,]  944  431
#>  [865,]  944  432
#>  [866,]  939  945
#>  [867,]  945  433
#>  [868,]  945  946
#>  [869,]  946  434
#>  [870,]  946  947
#>  [871,]  947  948
#>  [872,]  948  435
#>  [873,]  948  436
#>  [874,]  947  949
#>  [875,]  949  437
#>  [876,]  949  950
#>  [877,]  950  438
#>  [878,]  950  951
#>  [879,]  951  439
#>  [880,]  951  440
#>  [881,]  933  952
#>  [882,]  952  953
#>  [883,]  953  441
#>  [884,]  953  954
#>  [885,]  954  442
#>  [886,]  954  443
#>  [887,]  952  955
#>  [888,]  955  956
#>  [889,]  956  444
#>  [890,]  956  957
#>  [891,]  957  958
#>  [892,]  958  445
#>  [893,]  958  959
#>  [894,]  959  446
#>  [895,]  959  447
#>  [896,]  957  960
#>  [897,]  960  448
#>  [898,]  960  961
#>  [899,]  961  449
#>  [900,]  961  450
#>  [901,]  955  962
#>  [902,]  962  451
#>  [903,]  962  963
#>  [904,]  963  964
#>  [905,]  964  452
#>  [906,]  964  965
#>  [907,]  965  453
#>  [908,]  965  966
#>  [909,]  966  454
#>  [910,]  966  455
#>  [911,]  963  967
#>  [912,]  967  968
#>  [913,]  968  456
#>  [914,]  968  969
#>  [915,]  969  457
#>  [916,]  969  458
#>  [917,]  967  970
#>  [918,]  970  459
#>  [919,]  970  971
#>  [920,]  971  460
#>  [921,]  971  461
#>  [922,]  932  972
#>  [923,]  972  462
#>  [924,]  972  973
#>  [925,]  973  974
#>  [926,]  974  975
#>  [927,]  975  976
#>  [928,]  976  463
#>  [929,]  976  464
#>  [930,]  975  977
#>  [931,]  977  978
#>  [932,]  978  465
#>  [933,]  978  466
#>  [934,]  977  979
#>  [935,]  979  467
#>  [936,]  979  980
#>  [937,]  980  468
#>  [938,]  980  981
#>  [939,]  981  469
#>  [940,]  981  470
#>  [941,]  974  982
#>  [942,]  982  983
#>  [943,]  983  471
#>  [944,]  983  984
#>  [945,]  984  472
#>  [946,]  984  985
#>  [947,]  985  473
#>  [948,]  985  474
#>  [949,]  982  986
#>  [950,]  986  987
#>  [951,]  987  475
#>  [952,]  987  988
#>  [953,]  988  476
#>  [954,]  988  477
#>  [955,]  986  989
#>  [956,]  989  990
#>  [957,]  990  478
#>  [958,]  990  991
#>  [959,]  991  479
#>  [960,]  991  480
#>  [961,]  989  992
#>  [962,]  992  481
#>  [963,]  992  993
#>  [964,]  993  482
#>  [965,]  993  994
#>  [966,]  994  483
#>  [967,]  994  484
#>  [968,]  973  995
#>  [969,]  995  485
#>  [970,]  995  996
#>  [971,]  996  997
#>  [972,]  997  486
#>  [973,]  997  487
#>  [974,]  996  998
#>  [975,]  998  488
#>  [976,]  998  999
#>  [977,]  999 1000
#>  [978,] 1000 1001
#>  [979,] 1001  489
#>  [980,] 1001 1002
#>  [981,] 1002  490
#>  [982,] 1002  491
#>  [983,] 1000 1003
#>  [984,] 1003  492
#>  [985,] 1003 1004
#>  [986,] 1004 1005
#>  [987,] 1005  493
#>  [988,] 1005  494
#>  [989,] 1004 1006
#>  [990,] 1006 1007
#>  [991,] 1007  495
#>  [992,] 1007  496
#>  [993,] 1006 1008
#>  [994,] 1008  497
#>  [995,] 1008  498
#>  [996,]  999 1009
#>  [997,] 1009 1010
#>  [998,] 1010  499
#>  [999,] 1010 1011
#> [1000,] 1011  500
#> [1001,] 1011 1012
#> [1002,] 1012  501
#> [1003,] 1012  502
#> [1004,] 1009 1013
#> [1005,] 1013  503
#> [1006,] 1013 1014
#> [1007,] 1014  504
#> [1008,] 1014 1015
#> [1009,] 1015 1016
#> [1010,] 1016  505
#> [1011,] 1016  506
#> [1012,] 1015 1017
#> [1013,] 1017 1018
#> [1014,] 1018  507
#> [1015,] 1018  508
#> [1016,] 1017 1019
#> [1017,] 1019  509
#> [1018,] 1019  510
#> 
#> $edge.length
#>    [1] 8.035426e-03 4.009259e-02 3.362844e-02 8.035426e-03 2.022804e-02
#>    [6] 8.880840e-03 7.085848e-07 1.801549e-02 3.395366e-02 1.636396e-02
#>   [11] 2.711809e-02 2.253626e-02 1.522106e-02 1.917553e-02 3.504873e-03
#>   [16] 3.263658e-03 1.188447e-02 1.607634e-02 6.159892e-02 8.137776e-03
#>   [21] 3.183375e-02 1.682843e-02 7.005172e-03 1.676074e-02 3.806749e-03
#>   [26] 2.997489e-02 7.249246e-03 2.498857e-02 2.937883e-03 3.575905e-02
#>   [31] 3.535051e-03 2.345953e-02 9.658871e-03 9.047844e-03 8.932785e-03
#>   [36] 1.788039e-03 1.282331e-02 3.986246e-02 1.194906e-02 1.590474e-02
#>   [41] 4.537313e-02 5.321140e-03 1.823528e-02 5.532079e-03 3.459999e-02
#>   [46] 1.029555e-03 2.509621e-02 2.101002e-02 1.012876e-02 8.255296e-03
#>   [51] 1.561761e-02 1.024510e-02 9.808274e-02 1.008165e-02 3.052544e-03
#>   [56] 4.677525e-03 2.115559e-03 5.204426e-02 2.138133e-03 1.711925e-01
#>   [61] 3.935085e-03 2.961639e-02 3.533954e-02 1.710641e-02 3.370666e-03
#>   [66] 3.888387e-03 1.191943e-02 5.529678e-03 1.436714e-02 1.651331e-02
#>   [71] 2.705461e-02 1.309157e-02 5.731172e-02 8.338371e-03 1.766878e-02
#>   [76] 8.873915e-03 1.225312e-02 1.136802e-02 1.703329e-02 5.135252e-03
#>   [81] 1.237699e-02 1.772995e-02 1.046642e-02 3.062114e-03 1.548764e-02
#>   [86] 1.436431e-02 4.414238e-03 1.940605e-02 3.587769e-02 3.059545e-02
#>   [91] 3.552361e-02 1.256455e-02 4.905395e-03 4.161796e-03 8.150767e-03
#>   [96] 7.727008e-04 2.205107e-03 1.373115e-02 3.768393e-03 4.047170e-03
#>  [101] 7.268076e-03 2.377061e-02 5.651512e-03 2.533720e-03 1.455415e-02
#>  [106] 7.284218e-04 6.226336e-03 7.085848e-07 7.085848e-07 7.373415e-04
#>  [111] 2.196303e-03 4.053924e-03 4.167393e-03 7.816794e-03 2.175384e-02
#>  [116] 3.239160e-02 1.505292e-02 7.085848e-07 1.382285e-02 1.307633e-02
#>  [121] 2.567897e-02 1.207871e-02 1.053164e-02 3.166615e-03 4.225978e-03
#>  [126] 2.021017e-02 2.048564e-03 2.453640e-03 1.300326e-02 1.716726e-03
#>  [131] 4.906589e-03 7.831442e-03 2.832789e-03 1.043618e-02 3.738097e-02
#>  [136] 2.460126e-02 7.219234e-03 5.259747e-02 1.937757e-02 6.844655e-02
#>  [141] 6.203184e-03 1.173728e-02 2.150074e-02 2.354849e-02 1.600741e-02
#>  [146] 7.962633e-03 1.027362e-02 7.085848e-07 1.760611e-02 1.224087e-03
#>  [151] 3.800243e-03 5.085436e-03 7.102079e-03 1.140391e-02 8.630842e-03
#>  [156] 3.918699e-02 2.941855e-03 1.587738e-02 1.590703e-02 3.965877e-03
#>  [161] 5.921749e-02 6.794495e-03 1.785399e-03 1.047114e-02 7.085848e-07
#>  [166] 5.355837e-02 8.429108e-03 4.685591e-02 2.271520e-03 1.559243e-02
#>  [171] 5.038749e-02 2.555373e-02 1.887584e-02 1.518829e-02 1.896918e-02
#>  [176] 7.085848e-07 1.869201e-02 2.115265e-02 3.280065e-02 1.766151e-02
#>  [181] 2.653948e-02 6.068287e-03 8.570708e-03 1.899927e-02 9.851103e-03
#>  [186] 1.004105e-02 2.132079e-02 1.135432e-02 1.730698e-02 5.737558e-03
#>  [191] 5.993825e-03 3.215323e-03 9.489582e-03 4.397901e-02 6.781438e-03
#>  [196] 2.233158e-03 3.776972e-03 8.836507e-03 1.559238e-02 2.909034e-02
#>  [201] 1.864043e-02 1.748708e-02 1.059962e-02 2.861411e-03 9.390169e-04
#>  [206] 3.108475e-03 1.441067e-02 1.130029e-02 5.215559e-02 4.790619e-03
#>  [211] 6.345913e-04 1.698432e-03 4.541565e-03 1.973078e-02 2.196352e-02
#>  [216] 7.295308e-03 8.842118e-03 1.289925e-02 1.092495e-02 1.561116e-03
#>  [221] 3.181981e-02 6.586942e-04 7.085848e-07 1.119029e-02 8.170355e-03
#>  [226] 1.142162e-02 3.487893e-02 7.085848e-07 2.648518e-03 3.068780e-03
#>  [231] 2.387320e-02 1.273935e-02 2.143288e-02 3.112434e-02 7.085848e-07
#>  [236] 6.109249e-03 1.384891e-02 2.812922e-03 4.660207e-03 5.010224e-03
#>  [241] 2.453933e-02 1.651638e-02 1.943344e-03 1.597989e-02 1.862314e-02
#>  [246] 7.514722e-03 6.867320e-03 5.001598e-02 5.083922e-03 1.882011e-02
#>  [251] 1.580893e-03 5.201606e-03 1.064542e-03 2.390115e-03 1.573635e-02
#>  [256] 6.159009e-03 7.085848e-07 3.309455e-03 1.019717e-02 2.980094e-03
#>  [261] 7.085848e-07 7.085848e-07 3.860914e-02 2.463167e-03 2.464617e-03
#>  [266] 9.807741e-04 2.153113e-02 4.328222e-03 4.214043e-03 2.917108e-02
#>  [271] 1.445008e-02 7.552929e-03 1.227015e-02 2.158702e-03 3.424074e-03
#>  [276] 4.281895e-03 3.648645e-03 6.991253e-03 3.055729e-03 1.542014e-02
#>  [281] 5.132698e-03 4.750994e-03 1.131814e-02 7.085848e-07 3.869066e-02
#>  [286] 4.319138e-03 2.625468e-02 1.604332e-02 1.652808e-02 8.651163e-03
#>  [291] 4.996663e-03 1.304783e-02 6.049565e-03 6.082841e-03 1.731738e-02
#>  [296] 9.582062e-03 1.875985e-02 3.101731e-03 3.383852e-02 4.186548e-03
#>  [301] 1.146092e-03 2.955934e-02 1.320132e-02 2.639318e-02 1.351785e-02
#>  [306] 7.085848e-07 7.085848e-07 7.085848e-07 6.423223e-03 5.951659e-03
#>  [311] 2.281859e-02 1.208381e-03 3.283252e-02 2.335953e-02 2.381616e-03
#>  [316] 1.367892e-02 9.870760e-03 8.889860e-03 1.893028e-02 5.927800e-03
#>  [321] 2.134385e-02 7.085848e-07 2.666569e-02 2.968946e-02 7.947729e-03
#>  [326] 1.651392e-03 1.249678e-03 3.865624e-02 2.672850e-03 2.567447e-02
#>  [331] 2.860936e-02 5.200538e-03 4.668364e-03 1.781532e-02 3.577269e-03
#>  [336] 7.251906e-03 8.028196e-03 1.646062e-03 7.085848e-07 7.085848e-07
#>  [341] 3.927325e-03 8.480859e-03 6.056076e-03 1.841300e-03 8.295487e-04
#>  [346] 7.826039e-04 7.158814e-04 3.707568e-03 9.768438e-03 4.249502e-03
#>  [351] 7.085848e-07 7.348317e-03 4.358546e-03 5.971950e-03 1.458116e-03
#>  [356] 7.085848e-07 3.079206e-03 3.856124e-03 1.272690e-02 7.085848e-07
#>  [361] 2.416686e-03 5.048034e-03 2.925461e-03 2.925475e-03 4.172282e-03
#>  [366] 1.100966e-02 2.782694e-03 7.073536e-03 1.466371e-03 9.374979e-03
#>  [371] 9.681299e-03 7.085848e-07 1.519416e-03 6.896193e-03 8.225134e-04
#>  [376] 8.125060e-04 7.085848e-07 5.988672e-03 3.588312e-03 3.460973e-03
#>  [381] 2.752533e-02 7.640540e-03 7.085848e-07 7.085848e-07 1.651017e-03
#>  [386] 5.055038e-03 1.741173e-02 9.845938e-03 7.085848e-07 7.085848e-07
#>  [391] 1.571501e-03 1.446682e-02 2.626004e-03 1.627044e-02 7.519013e-04
#>  [396] 4.377728e-03 2.113955e-02 2.593769e-03 7.085848e-07 2.703375e-02
#>  [401] 3.522031e-02 6.394545e-03 3.768039e-03 6.363485e-03 1.688276e-03
#>  [406] 1.694933e-03 8.192864e-03 1.803830e-02 7.085848e-07 1.028044e-02
#>  [411] 2.685463e-02 1.060425e-02 8.103881e-04 4.502459e-03 3.640885e-03
#>  [416] 4.135469e-03 4.905194e-03 1.258398e-02 3.421898e-03 6.586773e-03
#>  [421] 7.085848e-07 8.053583e-03 1.308802e-03 8.135029e-03 1.828431e-03
#>  [426] 2.255645e-03 7.922336e-03 5.317068e-04 8.038783e-03 4.582472e-03
#>  [431] 7.729268e-04 1.214628e-02 1.450370e-02 1.787306e-03 3.701713e-03
#>  [436] 2.063123e-03 2.122448e-03 1.140345e-03 1.767978e-03 7.085848e-07
#>  [441] 9.690545e-03 7.085848e-07 6.462296e-03 1.037806e-02 3.233499e-03
#>  [446] 7.085848e-07 7.993098e-04 5.555695e-03 8.103633e-04 5.661940e-03
#>  [451] 3.368611e-04 3.828768e-03 7.279581e-04 1.275968e-03 1.517215e-03
#>  [456] 5.547882e-03 1.791285e-03 7.903532e-04 3.813386e-03 1.612963e-03
#>  [461] 2.178691e-02 7.085848e-07 7.787848e-04 2.430098e-03 1.106651e-03
#>  [466] 1.851449e-02 4.530983e-03 8.305496e-04 4.551938e-03 4.353991e-03
#>  [471] 9.621293e-03 7.085848e-07 7.255054e-03 9.259892e-04 4.835700e-03
#>  [476] 4.958559e-03 8.854962e-03 6.505757e-03 1.231949e-02 2.643935e-03
#>  [481] 1.756132e-02 1.446007e-02 1.185114e-03 1.644413e-03 1.507694e-02
#>  [486] 4.016740e-04 8.883443e-03 2.160049e-02 8.573250e-04 8.120340e-03
#>  [491] 7.085848e-07 2.056317e-03 1.765674e-03 7.085848e-07 3.348893e-03
#>  [496] 2.169954e-03 7.085848e-07 1.678161e-03 1.446878e-03 1.581256e-03
#>  [501] 3.379166e-03 2.160603e-03 1.422352e-03 7.085848e-07 3.669582e-03
#>  [506] 7.085848e-07 1.587284e-03 1.685288e-03 2.023673e-03 7.085848e-07
#>  [511] 8.819756e-04 7.085848e-07 2.176317e-03 3.360368e-03 7.085848e-07
#>  [516] 7.770098e-03 8.283878e-04 7.784595e-04 6.621717e-03 7.085848e-07
#>  [521] 4.402530e-03 8.014124e-04 2.192446e-03 7.393866e-04 8.376342e-04
#>  [526] 1.048021e-03 6.473868e-03 4.437709e-03 7.085848e-07 7.085848e-07
#>  [531] 9.370011e-04 3.112448e-03 1.457835e-02 2.280082e-03 2.625180e-03
#>  [536] 4.580187e-03 7.945719e-04 2.528598e-03 7.085848e-07 3.047672e-03
#>  [541] 7.085848e-07 1.445555e-03 1.463408e-03 3.425073e-03 7.085848e-07
#>  [546] 1.600677e-03 7.398699e-04 1.474027e-03 7.362372e-04 6.700967e-03
#>  [551] 7.085848e-07 8.416662e-04 3.536764e-02 1.061739e-02 5.371471e-03
#>  [556] 3.767208e-02 1.902905e-03 3.155914e-02 9.283271e-03 2.800354e-03
#>  [561] 2.111609e-03 2.196458e-03 2.202005e-03 3.968990e-03 4.179847e-03
#>  [566] 2.115035e-03 1.262514e-02 8.064353e-03 1.864562e-03 1.578342e-03
#>  [571] 7.690765e-04 9.389735e-03 6.014510e-03 3.966691e-03 1.191791e-02
#>  [576] 1.083669e-03 1.369696e-02 2.966783e-03 7.059460e-03 7.085848e-07
#>  [581] 2.736212e-02 9.260550e-04 4.881890e-03 1.501184e-03 1.599686e-02
#>  [586] 4.606906e-03 2.800602e-03 1.572111e-03 5.408565e-03 1.244069e-02
#>  [591] 1.138154e-02 7.085848e-07 1.252398e-02 1.716321e-03 1.940407e-02
#>  [596] 3.101903e-02 5.096520e-02 3.820518e-03 2.952933e-02 1.495585e-02
#>  [601] 4.906236e-03 7.085848e-07 5.430828e-03 8.600799e-03 5.106062e-03
#>  [606] 5.089325e-03 1.621773e-03 1.266649e-02 1.338771e-02 5.109999e-03
#>  [611] 2.182603e-02 3.384240e-03 9.755232e-03 5.875030e-03 8.350847e-04
#>  [616] 5.316149e-03 1.156728e-02 4.332484e-03 5.573315e-03 6.703482e-03
#>  [621] 2.130053e-03 5.882340e-03 8.990573e-03 2.613192e-02 6.388916e-03
#>  [626] 8.073544e-03 1.237965e-02 4.149340e-03 6.109370e-03 6.357573e-03
#>  [631] 7.645607e-04 5.999884e-03 6.462690e-03 7.085848e-07 1.096036e-02
#>  [636] 7.085848e-07 3.746880e-02 1.628139e-02 4.203531e-03 1.653534e-03
#>  [641] 3.976760e-03 6.179075e-02 4.210301e-02 7.085848e-07 2.869223e-02
#>  [646] 7.214371e-03 1.762207e-02 5.977163e-03 6.577968e-03 1.416854e-02
#>  [651] 1.092026e-02 2.104095e-03 1.275864e-03 3.829457e-03 1.427259e-02
#>  [656] 1.120660e-03 3.891638e-03 7.085848e-07 1.809276e-03 3.050277e-03
#>  [661] 1.960438e-03 5.992571e-03 1.472991e-02 1.076497e-02 4.048899e-03
#>  [666] 1.027648e-03 7.085848e-07 2.947371e-03 1.449438e-03 7.085848e-07
#>  [671] 7.085848e-07 7.579203e-03 7.085848e-07 6.195448e-03 7.879332e-04
#>  [676] 1.349165e-02 5.830686e-03 4.546097e-03 8.069505e-03 6.339558e-03
#>  [681] 3.047840e-03 3.684008e-03 1.529693e-03 2.216128e-03 4.286687e-03
#>  [686] 4.245837e-02 1.476662e-03 5.117715e-03 2.087639e-02 9.052073e-03
#>  [691] 3.039363e-03 9.074242e-03 1.703342e-03 3.301574e-03 8.041855e-03
#>  [696] 1.012905e-02 8.562925e-04 3.857445e-03 5.550378e-03 2.401379e-03
#>  [701] 1.336217e-02 9.837606e-03 4.404658e-03 6.412048e-03 8.878253e-03
#>  [706] 2.400646e-02 2.333687e-03 5.245767e-02 4.362142e-03 7.405588e-04
#>  [711] 8.097447e-03 8.254444e-04 9.926806e-03 5.184852e-03 1.501024e-02
#>  [716] 2.094842e-03 1.242651e-03 2.844290e-03 2.254942e-03 7.085848e-07
#>  [721] 1.460328e-02 9.956509e-04 2.577779e-03 1.174627e-02 7.113740e-04
#>  [726] 1.259470e-03 1.906344e-03 2.818743e-03 2.124099e-03 2.096742e-03
#>  [731] 7.078558e-04 5.717158e-03 7.085848e-07 4.360330e-04 5.551814e-03
#>  [736] 2.773655e-02 7.454558e-03 7.085848e-07 1.433831e-03 6.361694e-03
#>  [741] 1.725459e-02 9.148429e-03 1.436995e-02 5.795447e-03 4.145611e-03
#>  [746] 1.522270e-03 4.870136e-03 8.827563e-03 1.033050e-03 7.398348e-04
#>  [751] 1.417312e-02 2.757082e-03 1.216130e-03 7.085848e-07 1.346216e-02
#>  [756] 2.014939e-03 1.771235e-02 4.157066e-03 7.772567e-03 7.085848e-07
#>  [761] 7.085848e-07 1.407658e-03 7.085848e-07 1.021130e-03 5.240406e-03
#>  [766] 5.656892e-04 1.525336e-02 3.876103e-03 7.085848e-07 1.559759e-02
#>  [771] 8.519528e-03 2.077304e-03 4.447602e-03 2.435596e-03 2.246211e-03
#>  [776] 1.490590e-03 1.910232e-03 8.441112e-04 2.779685e-03 1.579941e-02
#>  [781] 3.030908e-03 1.480474e-03 1.564923e-03 7.868456e-04 3.112089e-03
#>  [786] 3.497685e-03 1.140562e-02 1.340509e-03 1.172596e-03 3.299868e-03
#>  [791] 7.362877e-03 2.400906e-03 1.623819e-02 3.378432e-03 7.085848e-07
#>  [796] 5.213761e-03 7.085848e-07 3.191650e-03 2.775136e-03 1.418766e-02
#>  [801] 4.174965e-02 2.209068e-03 6.121273e-03 4.360741e-02 5.700316e-02
#>  [806] 5.672951e-03 2.599482e-03 1.781967e-02 1.163604e-03 1.200705e-02
#>  [811] 1.273596e-02 3.155254e-03 7.085848e-07 7.085848e-07 8.763347e-04
#>  [816] 1.212596e-02 1.172467e-03 1.186436e-02 9.743986e-04 1.923284e-02
#>  [821] 8.568384e-04 6.358189e-03 5.961061e-03 1.213145e-02 7.755229e-03
#>  [826] 2.051308e-03 2.710419e-03 2.460856e-02 9.233620e-03 1.390873e-02
#>  [831] 2.597459e-04 1.022960e-02 4.584779e-03 2.220609e-03 2.415881e-03
#>  [836] 2.563789e-02 6.266087e-03 4.939583e-03 2.538677e-03 3.979690e-03
#>  [841] 9.440285e-03 1.302670e-03 5.161017e-03 1.129976e-04 1.346576e-02
#>  [846] 4.568760e-04 2.527489e-03 2.161837e-02 9.339914e-03 7.085848e-07
#>  [851] 1.069278e-03 2.248199e-03 1.101063e-03 9.153934e-04 7.181845e-03
#>  [856] 4.431229e-03 3.581661e-03 2.121248e-03 7.945147e-04 1.652382e-03
#>  [861] 1.415070e-02 6.784728e-03 3.448651e-03 2.281005e-03 8.035516e-03
#>  [866] 3.732273e-03 3.923237e-03 8.252646e-03 7.541208e-03 1.038244e-03
#>  [871] 4.977340e-03 3.822634e-03 2.975217e-03 7.315026e-04 7.337654e-03
#>  [876] 7.731573e-04 7.085848e-07 3.597985e-03 6.186511e-03 7.085848e-07
#>  [881] 8.950304e-04 1.031240e-02 1.175474e-02 3.001784e-03 2.299730e-02
#>  [886] 1.597533e-03 4.369055e-03 2.970846e-03 1.688660e-02 2.065480e-03
#>  [891] 5.120711e-03 1.107362e-03 1.156409e-03 7.085848e-07 2.371532e-03
#>  [896] 2.159003e-03 6.743833e-03 9.973233e-04 3.200683e-03 2.643002e-03
#>  [901] 4.345781e-03 1.572155e-02 9.908103e-04 2.242408e-02 1.858548e-02
#>  [906] 1.495275e-02 7.085848e-07 7.085848e-07 9.256399e-03 1.300391e-02
#>  [911] 4.533988e-03 3.099731e-03 3.703378e-03 1.750503e-03 2.699539e-03
#>  [916] 3.245517e-03 6.125781e-03 6.021033e-03 6.172100e-03 5.840047e-03
#>  [921] 2.536046e-03 2.190141e-03 2.192737e-02 4.508337e-03 2.000017e-03
#>  [926] 2.557309e-03 9.315422e-03 5.210062e-03 2.180467e-02 9.902914e-03
#>  [931] 7.085848e-07 7.085848e-07 4.438437e-03 2.081582e-03 2.337915e-03
#>  [936] 3.741507e-03 5.915871e-03 7.085848e-07 5.910653e-03 6.065071e-03
#>  [941] 3.856954e-03 1.395450e-03 1.536528e-02 1.690869e-03 7.488259e-03
#>  [946] 7.095588e-03 8.653127e-03 3.640362e-03 2.253299e-03 7.085848e-07
#>  [951] 1.554710e-02 3.531962e-03 2.165419e-02 1.564434e-02 1.828710e-03
#>  [956] 7.934260e-03 1.241051e-02 2.258958e-03 6.682940e-03 2.794756e-03
#>  [961] 7.085848e-07 1.577483e-02 2.394889e-03 1.976400e-02 2.761942e-03
#>  [966] 4.892397e-02 1.962326e-03 2.561422e-03 1.855289e-02 4.105072e-03
#>  [971] 7.085848e-07 9.679345e-03 2.005752e-02 2.529749e-03 3.779346e-03
#>  [976] 8.138741e-04 3.008653e-03 1.668463e-03 2.638335e-02 3.809465e-03
#>  [981] 7.085848e-07 3.837897e-03 8.334908e-04 1.476477e-02 8.743778e-04
#>  [986] 4.218314e-03 3.557093e-03 4.299889e-02 7.755493e-04 1.325415e-03
#>  [991] 1.301170e-03 4.550623e-03 3.355293e-03 2.442785e-03 3.425405e-03
#>  [996] 7.085848e-07 5.097122e-03 5.537954e-03 2.804232e-03 2.024693e-02
#> [1001] 2.684892e-03 7.085848e-07 7.085848e-07 8.886584e-04 9.431362e-03
#> [1006] 7.085848e-07 9.607115e-03 7.256989e-04 8.659938e-04 1.093041e-02
#> [1011] 3.112887e-03 6.125903e-03 6.862289e-03 7.085848e-07 7.516456e-04
#> [1016] 6.583965e-04 1.505483e-02 7.085848e-07
#> 
#> $Nnode
#> [1] 509
#> 
#> $tip.label
#>   [1] "Pararchidendron_pruinosum" "Paraserianthes_lophantha" 
#>   [3] "adinophylla"               "semicircinalis"           
#>   [5] "aphanoclada"               "inaequilatera"            
#>   [7] "marramamba"                "pickardii"                
#>   [9] "cuspidifolia"              "dempsteri"                
#>  [11] "pyrifolia"                 "strongylophylla"          
#>  [13] "congesta"                  "victoriae"                
#>  [15] "alexandri"                 "synchronicia"             
#>  [17] "ryaniana"                  "chartacea"                
#>  [19] "bakeri"                    "saligna"                  
#>  [21] "gregorii"                  "longispinea"              
#>  [23] "aculeatissima"             "carnosula"                
#>  [25] "complanata"                "hylonoma"                 
#>  [27] "multisiliqua"              "simsii"                   
#>  [29] "obtecta"                   "siculiformis"             
#>  [31] "maitlandii"                "estrophiolata"            
#>  [33] "excelsa"                   "continua"                 
#>  [35] "spinescens"                "amblygona"                
#>  [37] "pravifolia"                "basedowii"                
#>  [39] "oxyclada"                  "acanthoclada"             
#>  [41] "acutata"                   "woodmaniorum"             
#>  [43] "pterocaulon"               "aphylla"                  
#>  [45] "ashbyae"                   "ampliceps"                
#>  [47] "salicina"                  "telmica"                  
#>  [49] "cupularis"                 "bidentata"                
#>  [51] "tysonii"                   "bivenosa"                 
#>  [53] "ligulata"                  "didyma"                   
#>  [55] "rostellifera"              "xanthina"                 
#>  [57] "spathulifolia"             "scirpifolia"              
#>  [59] "blakelyi"                  "porcata"                  
#>  [61] "hypermeces"                "anasilla"                 
#>  [63] "adoxa"                     "lycopodiifolia"           
#>  [65] "asperulacea"               "perryi"                   
#>  [67] "extensa"                   "gonophylla"               
#>  [69] "alata"                     "applanata"                
#>  [71] "shuttleworthii"            "pentadenia"               
#>  [73] "empelioclada"              "browniana"                
#>  [75] "nigricans"                 "gilbertii"                
#>  [77] "leioderma"                 "fagonioides"              
#>  [79] "epacantha"                 "pulchella"                
#>  [81] "guinetii"                  "drummondii"               
#>  [83] "jacksonioides"             "pygmaea"                  
#>  [85] "incrassata"                "phlebopetala"             
#>  [87] "biflora"                   "clydonophora"             
#>  [89] "heterochroa"               "celastrifolia"            
#>  [91] "myrtifolia"                "gelasina"                 
#>  [93] "subrigida"                 "praelongata"              
#>  [95] "murrayana"                 "pachyacra"                
#>  [97] "imbricata"                 "triquetra"                
#>  [99] "acinacea"                  "iteaphylla"               
#> [101] "suaveolens"                "anthochaera"              
#> [103] "prainii"                   "camptoclada"              
#> [105] "subsessilis"               "diallaga"                 
#> [107] "ixodes"                    "calantha"                 
#> [109] "melleodora"                "jensenii"                 
#> [111] "sabulosa"                  "dictyophleba"             
#> [113] "acuaria"                   "benthamii"                
#> [115] "auronitens"                "cavealis"                 
#> [117] "latipes"                   "cochlearis"               
#> [119] "adnata"                    "acanthaster"              
#> [121] "tetragonophylla"           "curvata"                  
#> [123] "declinata"                 "leptoneura"               
#> [125] "campylophylla"             "aulacophylla"             
#> [127] "brachyphylla"              "fragilis"                 
#> [129] "heteroclita"               "assimilis"                
#> [131] "stanleyi"                  "peuce"                    
#> [133] "saxatilis"                 "hystrix"                  
#> [135] "rhamphophylla"             "glaucoptera"              
#> [137] "bifaria"                   "excentrica"               
#> [139] "halliana"                  "acoma"                    
#> [141] "glaucissima"               "pruinocarpa"              
#> [143] "latescens"                 "platycarpa"               
#> [145] "dunnii"                    "retivenea"                
#> [147] "auricoma"                  "andrewsii"                
#> [149] "baxteri"                   "huegelii"                 
#> [151] "semitrullata"              "kybeanensis"              
#> [153] "hastulata"                 "aemula"                   
#> [155] "cedroides"                 "ataxiphylla"              
#> [157] "costata"                   "bracteolata"              
#> [159] "carneorum"                 "atrox"                    
#> [161] "spectabilis"               "validinervia"             
#> [163] "terminalis"                "elata"                    
#> [165] "latisepala"                "gladiiformis"             
#> [167] "beckleri"                  "pycnantha"                
#> [169] "pedina"                    "hakeoides"                
#> [171] "brachybotrya"              "argyrophylla"             
#> [173] "gillii"                    "notabilis"                
#> [175] "euthycarpa"                "subulata"                 
#> [177] "wattsiana"                 "hemiteles"                
#> [179] "jennerae"                  "calamifolia"              
#> [181] "araneosa"                  "alcockii"                 
#> [183] "anceps"                    "retinodes"                
#> [185] "nematophylla"              "aestivalis"               
#> [187] "meisneri"                  "rivalis"                  
#> [189] "harveyi"                   "amblyophylla"             
#> [191] "microbotrya"               "bartleana"                
#> [193] "macradenia"                "falcata"                  
#> [195] "bancroftiorum"             "decora"                   
#> [197] "rubida"                    "semilunata"               
#> [199] "beadleana"                 "pubicosta"                
#> [201] "tindaleae"                 "conferta"                 
#> [203] "storyi"                    "angusta"                  
#> [205] "anaticeps"                 "chrysotricha"             
#> [207] "linearifolia"              "adunca"                   
#> [209] "neriifolia"                "ingramii"                 
#> [211] "pustula"                   "fimbriata"                
#> [213] "prominens"                 "pravissima"               
#> [215] "boormanii"                 "leucolobia"               
#> [217] "clunies-rossiae"           "betchei"                  
#> [219] "perangusta"                "dorothea"                 
#> [221] "uncinata"                  "aureocrinita"             
#> [223] "falciformis"               "obliquinervia"            
#> [225] "acrionastes"               "penninervis"              
#> [227] "binata"                    "binervata"                
#> [229] "mabellae"                  "caesiella"                
#> [231] "saliciformis"              "covenyi"                  
#> [233] "macnuttiana"               "vestita"                  
#> [235] "barringtonensis"           "cremiflora"               
#> [237] "linifolia"                 "cultriformis"             
#> [239] "obtusata"                  "gittinsii"                
#> [241] "mitchellii"                "chinchillensis"           
#> [243] "attenuata"                 "buxifolia"                
#> [245] "debilis"                   "polybotrya"               
#> [247] "cardiophylla"              "dealbata"                 
#> [249] "baileyana"                 "nanodealbata"             
#> [251] "jonesii"                   "trachyphloia"             
#> [253] "blayana"                   "silvestris"               
#> [255] "decurrens"                 "glaucocarpa"              
#> [257] "filicifolia"               "leucoclada"               
#> [259] "muelleriana"               "oshanesii"                
#> [261] "schinoides"                "pruinosa"                 
#> [263] "fulva"                     "mollifolia"               
#> [265] "cangaiensis"               "dangarensis"              
#> [267] "jucunda"                   "podalyriifolia"           
#> [269] "loroloba"                  "pedleyi"                  
#> [271] "parramattensis"            "olsenii"                  
#> [273] "deanei"                    "irrorata"                 
#> [275] "parvipinnula"              "mearnsii"                 
#> [277] "constablei"                "wanyu"                    
#> [279] "cuthbertsonii"             "oldfieldii"               
#> [281] "jibberdingensis"           "karina"                   
#> [283] "triptera"                  "rhigiophylla"             
#> [285] "alpina"                    "courtii"                  
#> [287] "phlebophylla"              "obtusifolia"              
#> [289] "longissima"                "axillaris"                
#> [291] "mucronata"                 "floribunda"               
#> [293] "longifolia"                "masliniana"               
#> [295] "patagiata"                 "havilandiorum"            
#> [297] "colletioides"              "delphina"                 
#> [299] "simulans"                  "argutifolia"              
#> [301] "nyssophylla"               "mackeyana"                
#> [303] "inceana"                   "lineolata"                
#> [305] "baueri"                    "oswaldii"                 
#> [307] "calcicola"                 "rigens"                   
#> [309] "enterocarpa"               "hexaneura"                
#> [311] "montana"                   "dawsonii"                 
#> [313] "verricula"                 "wilhelmiana"              
#> [315] "ixiophylla"                "dictyoneura"              
#> [317] "brachypoda"                "auratiflora"              
#> [319] "consobrina"                "stenophylla"              
#> [321] "cyclops"                   "brachyclada"              
#> [323] "redolens"                  "georginae"                
#> [325] "harpophylla"               "argyrodendron"            
#> [327] "sclerophylla"              "ancistrophylla"           
#> [329] "amyctica"                  "enervia"                  
#> [331] "papyrocarpa"               "maranoensis"              
#> [333] "microsperma"               "melvillei"                
#> [335] "omalophylla"               "loderi"                   
#> [337] "maconochieana"             "tephrina"                 
#> [339] "pendula"                   "cana"                     
#> [341] "sibilans"                  "latzii"                   
#> [343] "dolichophylla"             "barattensis"              
#> [345] "spongolitica"              "crassiuscula"             
#> [347] "baeuerlenii"               "venulosa"                 
#> [349] "elongata"                  "genistifolia"             
#> [351] "tessellata"                "melanoxylon"              
#> [353] "implexa"                   "viscidula"                
#> [355] "flexifolia"                "errabunda"                
#> [357] "menzelii"                  "lineata"                  
#> [359] "rhetinocarpa"              "paradoxa"                 
#> [361] "aspera"                    "dodonaeifolia"            
#> [363] "howittii"                  "cognata"                  
#> [365] "verniciflua"               "gnidium"                  
#> [367] "ausfeldii"                 "sericophylla"             
#> [369] "coriacea"                  "orites"                   
#> [371] "pachycarpa"                "xiphophylla"              
#> [373] "aciphylla"                 "sibina"                   
#> [375] "beauverdiana"              "exilis"                   
#> [377] "hamersleyensis"            "lasiocalyx"               
#> [379] "olgana"                    "tarculensis"              
#> [381] "umbraculiformis"           "citrinoviridis"           
#> [383] "jamesiana"                 "bromilowiana"             
#> [385] "ammobia"                   "petraea"                  
#> [387] "desmondii"                 "undoolyana"               
#> [389] "curranii"                  "repanda"                  
#> [391] "oncinophylla"              "multispicata"             
#> [393] "armitii"                   "plectocarpa"              
#> [395] "delibrata"                 "gracillima"               
#> [397] "trachycarpa"               "chisholmii"               
#> [399] "lysiphloia"                "eriopoda"                 
#> [401] "hemsleyi"                  "blakei"                   
#> [403] "difficilis"                "mountfordiae"             
#> [405] "tumida"                    "areolata"                 
#> [407] "retinervis"                "julifera"                 
#> [409] "stipuligera"               "brassii"                  
#> [411] "proiantha"                 "conspersa"                
#> [413] "torulosa"                  "drepanophylla"            
#> [415] "neurophylla"               "sessilispica"             
#> [417] "denticulosa"               "nuperrima"                
#> [419] "burkittii"                 "acuminata"                
#> [421] "anfractuosa"               "anastema"                 
#> [423] "sparsiflora"               "hopperiana"               
#> [425] "aprica"                    "arcuatilis"               
#> [427] "subtessarogona"            "balsamea"                 
#> [429] "coolgardiensis"            "sulcaticaulis"            
#> [431] "ramulosa"                  "brachystachya"            
#> [433] "catenulata"                "distans"                  
#> [435] "craspedocarpa"             "aneura"                   
#> [437] "clelandii"                 "minyura"                  
#> [439] "ayersiana"                 "paraneura"                
#> [441] "atkinsiana"                "macdonnellensis"          
#> [443] "adsurgens"                 "umbellata"                
#> [445] "stowardii"                 "aprepta"                  
#> [447] "kempeana"                  "cyperophylla"             
#> [449] "rhodophloia"               "grasbyi"                  
#> [451] "tenuissima"                "doratoxylon"              
#> [453] "pubifolia"                 "diphylla"                 
#> [455] "pycnostachya"              "caroleae"                 
#> [457] "striatifolia"              "tenuinervis"              
#> [459] "cheelii"                   "binervia"                 
#> [461] "bulgaensis"                "monticola"                
#> [463] "translucens"               "oncinocarpa"              
#> [465] "crassicarpa"               "aulacocarpa"              
#> [467] "lamprocarpa"               "celsa"                    
#> [469] "midgleyi"                  "disparrima"               
#> [471] "yirrkallensis"             "ptychophylla"             
#> [473] "abbreviata"                "hilliana"                 
#> [475] "producta"                  "stigmatophylla"           
#> [477] "argyraea"                  "ancistrocarpa"            
#> [479] "arrecta"                   "arida"                    
#> [481] "orthocarpa"                "tenuispica"               
#> [483] "gonocarpa"                 "drepanocarpa"             
#> [485] "tropica"                   "acradenia"                
#> [487] "dimidiata"                 "hammondii"                
#> [489] "auriculiformis"            "mangium"                  
#> [491] "spirorbis"                 "leptocarpa"               
#> [493] "faucium"                   "longispicata"             
#> [495] "concurrens"                "leiocalyx"                
#> [497] "crassa"                    "grandifolia"              
#> [499] "gardneri"                  "brockii"                  
#> [501] "neurocarpa"                "holosericea"              
#> [503] "gonoclada"                 "fecunda"                  
#> [505] "sericoflora"               "pellita"                  
#> [507] "colei"                     "thomsonii"                
#> [509] "cowleana"                  "elachantha"               
#> 
#> attr(,"class")
#> [1] "phylo"
#> attr(,"order")
#> [1] "cladewise"

# Example community
acacia$comm[1:5,1:5]

#>                     abbreviata acanthaster acanthoclada acinacea aciphylla
#> '-1025000:-1825000'          0           0            0        0         0
#> '-1025000:-1875000'          0           0            0        0         0
#> '-1025000:-1925000'          0           0            0        0         0
#> '-1025000:-1975000'          0           0            0        0         0
#> '-1025000:-2025000'          0           0            0        0         0