Calculate diversity metrics for a single random community

The independent swap method of Gotelli (2000) is used, which randomizes the community matrix while maintaining species occurrence frequency and sample species richness.

calc_biodiv_random(
  comm,
  phy,
  phy_alt,
  null_model = c("frequency", "richness", "independentswap", "trialswap"),
  n_iterations = 1000,
  metrics
)

Arguments

comm
Input community matrix in data.frame format (communities as rows, species as columns, with row names and column names)

phy
Input phylogeny with total branch length scaled to 1

phy_alt
Alternative phylogeny where all branches are of equal length, scaled to 1

n_iterations
Number of iterations to use when shuffling random community

metrics
Character vector; names of metrics to calculate

Value

List of vectors. Each vector is a biodiversity metric measured on the random community, in the same order as the rows in the input community.

Details

For description of metrics available, see run_ses_analysis()

Examples

if (FALSE) {
library(picante)
data(phylocom)
phy <- phylocom$phy
comm <- phylocom$sample
subsetted_data <- picante::match.phylo.comm(phy = phy, comm = comm)
phy <- subsetted_data[["phy"]]
comm <- subsetted_data[["comm"]]
phy_alt <- phy
phy_alt$edge.length <- rep(x = 1, times = length(phy_alt$edge.length))
phy_alt$edge.length <- phy_alt$edge.length / sum(phy_alt$edge.length)
phy$edge.length <- phy$edge.length / sum(phy$edge.length)
calc_biodiv_random(comm, phy, phy_alt, "independentswap", metrics = "pd")
}