Test the number of iterations needed to generate a random community that is sufficiently different from the original community.

A single random community will be created for each value in n_iter_test, using that value for the n_iterations argument of cpr_rand_comm(). The Euclidean distance is then calculated between the original community and each random community.

cpr_iter_test(
  comm,
  null_model = "curveball",
  binary = NULL,
  thin = 1,
  seed = NULL,
  n_iter_test,
  time_digits = 5,
  time_units = "secs"
)

Arguments

comm

Dataframe or matrix; input community data with sites (communities) as rows and species as columns. Values of each cell are the presence/absence (0 or 1) or number of individuals (abundance) of each species in each site.

null_model

Character vector of length 1 or object of class commsim; either the name of the model to use for generating random communities (null model), or a custom null model. For full list of available predefined null models, see the help file of vegan::commsim(), or run vegan::make.commsim(). An object of class commsim can be generated with vegan::commsim().

binary

Logical vector of length 1; is null model a binary model (produces random community data matrix with 1s and 0s) or not? If NULL (default), will attempt to determine automatically from name of null model (null_model).

thin

Numeric vector of length 1; thinning parameter used by some null models in vegan (e.g., quasiswap); ignored for other models.

seed

Integer vector of length 1 or NULL; random seed that will be used in a call to set.seed() before randomizing the matrix. Default (NULL) will not change the random generator state.

n_iter_test

Numeric vector; number of iterations to test (see Details and Examples).

time_digits

Numeric vector of length 1; number of digits to include when recording elapsed time to calculate each random community. Default 5.

time_units

Character vector of length 1; units to use when recording elapsed time to calculate each random community. Must be a valid value for units argument of base::difftime(). Default seconds (`"secs"``).

Value

Tibble (dataframe) with the following columns:

n_iter: Number of iterations used to generate random community

dist: Euclidean distance between original community and random community

time: Amount of time elapsed when generating random community

Details

The user should verify that distances increase with increasing iterations until an approximate maximum is reached. Of course, this only makes sense for randomization algorithms that use iterations.

The number of iterations to test (n_iter_test) is left to the user. A good rule of thumb is to test successive powers of 10. In this case, no more than four or five values is usually needed before distances stabilize (see Examples). Generally, fewer iterations are needed for smaller datasets.

Examples

data(biod_example)
cpr_iter_test(
  comm = biod_example$comm,
  null_model = "swap",
  n_iter_test = c(1, 10, 1000, 10000),
  seed = 123
)
#> # A tibble: 4 × 3
#>   n_iter  dist    time
#>    <int> <dbl>   <dbl>
#> 1      1  2    0.0014 
#> 2     10  6.32 0.00251
#> 3   1000 23.6  0.00909
#> 4  10000 23.8  0.0679