Run a randomization analysis for one or more biodiversity metrics

The observed value of the biodiversity metric(s) will be calculated, then compared against a set of random communities. Various statistics are calculated from the comparison (see Value below).

cpr_rand_test(
  comm,
  phy,
  null_model = "independentswap",
  n_reps = 100,
  n_iterations = 10000,
  metrics = c("pd", "rpd", "pe", "rpe"),
  site_col = "site",
  tbl_out = tibble::is_tibble(comm)
)

Arguments

comm
Dataframe, tibble, or matrix; input community data with sites (communities) as rows and species as columns. Values of each cell are the presence/absence (0 or 1) or number of individuals (abundance) of each species in each site.

phy
List of class phylo; input phylogeny.

null_model
Character vector of length 1; name of null model to use. Must choose from frequency, richness, independentswap, or trialswap (case-sensitive). For details, see picante::randomizeMatrix().

n_reps
Numeric vector of length 1; number of random communities to replicate.

n_iterations
Numeric vector of length 1; number of iterations to use when swapping occurrences to generate each random community; only used if null_model is independentswap or trialswap.

metrics
Character vector; names of biodiversity metrics to calculate. May include one or more of: pd, rpd, pe, rpe (case-sensitive).

site_col
Character vector of length 1; name of column in comm that contains the site names; only used if comm is a tibble (object of class tbl_df).

tbl_out
Logical vector of length 1; should the output be returned as a tibble? If FALSE, will return a dataframe. Defaults to TRUE if comm is a tibble.

Source

Faith DP (1992) Conservation evaluation and phylogenetic diversity. Biological Conservation, 61:1–10. doi: 10.1016/0006-3207(92)91201-3

Gotelli, N.J. (2000) Null Model Analysis of Species Co-Occurrence Patterns. Ecology, 81: 2606-2621. doi: 10.1890/0012-9658(2000)081[2606:NMAOSC]2.0.CO;2

Rosauer, D., Laffan, S.W., Crisp, M.D., Donnellan, S.C. and Cook, L.G. (2009) Phylogenetic endemism: a new approach for identifying geographical concentrations of evolutionary history. Molecular Ecology, 18: 4061-4072. doi: 10.1111/j.1365-294X.2009.04311.x

Mishler, B., Knerr, N., González-Orozco, C. et al. (2014) Phylogenetic measures of biodiversity and neo- and paleo-endemism in Australian Acacia. Nat Commun, 5: 4473. doi: 10.1038/ncomms5473

Value

Dataframe. For each of the biodiversity metrics, the following 9 columns will be produced:

*_obs: Observed value

*_obs_c_lower: Count of times observed value was lower than random values

*_obs_c_upper: Count of times observed value was higher than random values

*_obs_p_lower: Percentage of times observed value was lower than random values

*_obs_p_upper: Percentage of times observed value was higher than random values

*_obs_q: Count of the non-NA random values used for comparison

*_obs_z: Standard effect size (z-score)

*_rand_mean: Mean of the random values

*_rand_sd: Standard deviation of the random values

So if you included pd in metrics, the output columns would include pd_obs, pd_obs_c_lower, etc...

Details

The biodiversity metrics available for analysis include:

pd: Phylogenetic diversity (Faith 1992)

rpd: Relative phylogenetic diversity (Mishler et al 2014)

pe: Phylogenetic endemism (Rosauer et al 2009)

rpe: Relative phylogenetic endemism (Mishler et al 2014)

The default method for generating random communities is the independent swap method of Gotelli (2000), which randomizes the community matrix while maintaining species occurrence frequency and sample species richness.

A minimum of 5 species and sites are required as input; fewer than that may cause the default method for generating random communities (independent swap) to enter an infinite loop. Besides, inferences on very small numbers of species and/or sites is not recommended generally.

The following rules apply to comm input:

If dataframe or matrix, must include row names (site names) and column names (species names).

If tibble, a single column (default, site) must be included with site names, and other columns must correspond to species names.

Column names cannot start with a number and must be unique.

Row names (site names) must be unique.

Values (other than site names) should only include integers >= 0; non-integer input will be converted to integer.

The results are identical regardless of whether the input for comm is abundance or presence-absence data (i.e., abundance weighting is not used).

Examples

library(picante)
data(phylocom)
cpr_rand_test(phylocom$sample, phylocom$phy, metrics = "pd")

#> [1] "Dropping tips from the tree because they are not present in the community data:"
#> [1] "sp16" "sp23" "sp27" "sp28" "sp30" "sp31" "sp32"

#>            pd_obs pd_rand_mean pd_rand_sd  pd_obs_z pd_obs_c_upper
#> clump1  0.3018868    0.4733962 0.03023331 -5.672862              0
#> clump2a 0.3207547    0.4669811 0.03396173 -4.305623              0
#> clump2b 0.3396226    0.4694340 0.03650361 -3.556123              0
#> clump4  0.4150943    0.4684906 0.03392784 -1.573818              2
#> even    0.5660377    0.4649057 0.03376475  2.995197            100
#> random  0.5094340    0.4662264 0.02953541  1.462907             87
#>         pd_obs_c_lower pd_obs_q pd_obs_p_upper pd_obs_p_lower
#> clump1             100      100           0.00           1.00
#> clump2a            100      100           0.00           1.00
#> clump2b            100      100           0.00           1.00
#> clump4              91      100           0.02           0.91
#> even                 0      100           1.00           0.00
#> random               2      100           0.87           0.02