Classify phylogenetic endemism

Given the results of cpr_rand_test(), classifies phylogenetic endemism according to CANAPE scheme of Mishler et al. 2014.

cpr_classify_endem(df)

Arguments

df

Input data frame. Must have the following columns:

pe_obs_p_upper: Upper p-value comparing observed phylogenetic endemism to random values

pe_alt_obs_p_upper: Upper p-value comparing observed phylogenetic endemism on alternate tree to random values

rpe_obs_p_upper: Upper p-value comparing observed relative phylogenetic endemism to random values

Value

Object of class data.frame with column endem_type (character) added. Values of endem_type type include paleo (paleoendemic), neo(neoendemic), not significant (what it says), mixed (mixed endemism), and super (super-endemic; both pe_obs and pe_obs_alt are highly significant).

Details

For a summary of the classification scheme, see: http://biodiverse-analysis-software.blogspot.com/2014/11/canape-categorical-analysis-of-palaeo.html

References

Mishler, B., Knerr, N., González-Orozco, C. et al. (2014) Phylogenetic measures of biodiversity and neo- and paleo-endemism in Australian Acacia. Nat Commun, 5: 4473. doi: 10.1038/ncomms5473

Examples

set.seed(12345)
data(phylocom)
rand_test <- cpr_rand_test(
  phylocom$comm, phylocom$phy,
  null_model = "curveball", metrics = c("pe", "rpe")
)
#> Warning: Dropping tips from the tree because they are not present in the community data: 
#>  sp16, sp23, sp27, sp28, sp30, sp31, sp32
cpr_classify_endem(rand_test)
#>            pe_obs pe_rand_mean pe_rand_sd   pe_obs_z pe_obs_c_upper
#> clump1  0.1333333    0.1625283 0.02725696 -1.0711015             12
#> clump2a 0.1081761    0.1660660 0.02501413 -2.3142896              1
#> clump2b 0.1286164    0.1705660 0.02610899 -1.6067142              4
#> clump4  0.1411950    0.1655063 0.02837949 -0.8566512             20
#> even    0.2506289    0.1673648 0.02596072  3.2073123            100
#> random  0.2380503    0.1679686 0.02669970  2.6248141             99
#>         pe_obs_c_lower pe_obs_q pe_obs_p_upper pe_obs_p_lower pe_alt_obs
#> clump1              88      100           0.12           0.88  0.1472222
#> clump2a             99      100           0.01           0.99  0.1194444
#> clump2b             96      100           0.04           0.96  0.1420139
#> clump4              80      100           0.20           0.80  0.1454861
#> even                 0      100           1.00           0.00  0.2246528
#> random               1      100           0.99           0.01  0.2211806
#>         pe_alt_rand_mean pe_alt_rand_sd pe_alt_obs_z pe_alt_obs_c_upper
#> clump1         0.1633125     0.02269566   -0.7089583                 24
#> clump2a        0.1647882     0.02157452   -2.1017274                  1
#> clump2b        0.1703472     0.01992098   -1.4222859                  9
#> clump4         0.1659410     0.02354066   -0.8689163                 19
#> even           0.1677500     0.02116501    2.6885302                 99
#> random         0.1678611     0.02173363    2.4533153                 99
#>         pe_alt_obs_c_lower pe_alt_obs_q pe_alt_obs_p_upper pe_alt_obs_p_lower
#> clump1                  76          100               0.24               0.76
#> clump2a                 99          100               0.01               0.99
#> clump2b                 91          100               0.09               0.91
#> clump4                  81          100               0.19               0.81
#> even                     1          100               0.99               0.01
#> random                   1          100               0.99               0.01
#>           rpe_obs rpe_rand_mean rpe_rand_sd  rpe_obs_z rpe_obs_c_upper
#> clump1  0.9056604     0.9951685  0.09733498 -0.9195887               5
#> clump2a 0.9056604     1.0094275  0.10097860 -1.0276146               2
#> clump2b 0.9056604     1.0010135  0.09588822 -0.9944192              12
#> clump4  0.9705048     0.9977641  0.10357482 -0.2631844              56
#> even    1.1156280     0.9978474  0.09522202  1.2369050              87
#> random  1.0762714     1.0015190  0.10555106  0.7082111              79
#>         rpe_obs_c_lower rpe_obs_q rpe_obs_p_upper rpe_obs_p_lower
#> clump1               86       100            0.05            0.86
#> clump2a              93       100            0.02            0.93
#> clump2b              78       100            0.12            0.78
#> clump4               44       100            0.56            0.44
#> even                 13       100            0.87            0.13
#> random               21       100            0.79            0.21
#>              endem_type
#> clump1  not significant
#> clump2a not significant
#> clump2b not significant
#> clump4  not significant
#> even              mixed
#> random            mixed