Classify phylogenetic endemism

Given the results of cpr_rand_test(), classifies phylogenetic endemism according to CANAPE scheme of Mishler et al. 2014.

cpr_classify_endem(df)

Arguments

df
Input data frame. Must have the following columns:

pe_obs_p_upper: Upper p-value comparing observed phylogenetic endemism to random values

pe_alt_obs_p_upper: Upper p-value comparing observed phylogenetic endemism on alternate tree to random values

rpe_obs_p_upper: Upper p-value comparing observed relative phylogenetic endemism to random values

Value

Dataframe with column endem_type (character) added. Values of endem_type type include paleo (paleoendemic), neo (neoendemic), not significant (what it says), mixed (mixed endemism), and super (super-endemic; both pe_obs and pe_obs_alt are highly significant).

Details

For a summary of the classification scheme, see: http://biodiverse-analysis-software.blogspot.com/2014/11/canape-categorical-analysis-of-palaeo.html

References

Mishler, B., Knerr, N., González-Orozco, C. et al. (2014) Phylogenetic measures of biodiversity and neo- and paleo-endemism in Australian Acacia. Nat Commun, 5: 4473. doi: 10.1038/ncomms5473

Examples

data(phylocom)
rand_test <- cpr_rand_test(
  phylocom$sample, phylocom$phy,
  null_model = "curveball", metrics = c("pe", "rpe"))

#> [1] "Dropping tips from the tree because they are not present in the community data:"
#> [1] "sp16" "sp23" "sp27" "sp28" "sp30" "sp31" "sp32"

cpr_classify_endem(rand_test)

#>            pe_obs pe_rand_mean pe_rand_sd  pe_obs_z pe_obs_c_upper
#> clump1  0.1333333    0.1636887 0.02989514 -1.015394             13
#> clump2a 0.1081761    0.1666384 0.02960449 -1.974777              0
#> clump2b 0.1286164    0.1685943 0.02787814 -1.434026              8
#> clump4  0.1411950    0.1686289 0.02810313 -0.976189             14
#> even    0.2506289    0.1691195 0.02883897  2.826364             99
#> random  0.2380503    0.1633302 0.02851492  2.620387             97
#>         pe_obs_c_lower pe_obs_q pe_obs_p_upper pe_obs_p_lower pe_alt_obs
#> clump1              87      100           0.13           0.87  0.1472222
#> clump2a            100      100           0.00           1.00  0.1194444
#> clump2b             92      100           0.08           0.92  0.1420139
#> clump4              85      100           0.14           0.85  0.1454861
#> even                 1      100           0.99           0.01  0.2246528
#> random               3      100           0.97           0.03  0.2211806
#>         pe_alt_rand_mean pe_alt_rand_sd pe_alt_obs_z pe_alt_obs_c_upper
#> clump1         0.1660868     0.02502824   -0.7537319                 22
#> clump2a        0.1663924     0.02289137   -2.0509000                  0
#> clump2b        0.1666424     0.02349240   -1.0483590                 14
#> clump4         0.1669931     0.02007850   -1.0711431                 15
#> even           0.1710764     0.02317284    2.3120343                 99
#> random         0.1628090     0.02244057    2.6011608                 98
#>         pe_alt_obs_c_lower pe_alt_obs_q pe_alt_obs_p_upper pe_alt_obs_p_lower
#> clump1                  78          100               0.22               0.78
#> clump2a                100          100               0.00               1.00
#> clump2b                 86          100               0.14               0.86
#> clump4                  85          100               0.15               0.85
#> even                     1          100               0.99               0.01
#> random                   2          100               0.98               0.02
#>           rpe_obs rpe_rand_mean rpe_rand_sd  rpe_obs_z rpe_obs_c_upper
#> clump1  0.9056604     0.9842501  0.09224833 -0.8519367               6
#> clump2a 0.9056604     1.0004746  0.10465079 -0.9060054               6
#> clump2b 0.9056604     1.0124784  0.10129927 -1.0544798               9
#> clump4  0.9705048     1.0088078  0.11003457 -0.3481000              52
#> even    1.1156280     0.9880428  0.09659330  1.3208493              85
#> random  1.0762714     1.0030150  0.10592915  0.6915609              79
#>         rpe_obs_c_lower rpe_obs_q rpe_obs_p_upper rpe_obs_p_lower
#> clump1               84       100            0.06            0.84
#> clump2a              89       100            0.06            0.89
#> clump2b              84       100            0.09            0.84
#> clump4               48       100            0.52            0.48
#> even                 15       100            0.85            0.15
#> random               21       100            0.79            0.21
#>              endem_type
#> clump1  not significant
#> clump2a not significant
#> clump2b not significant
#> clump4  not significant
#> even              mixed
#> random            mixed