Classify phylogenetic endemism

Given the results of cpr_rand_test(), classifies phylogenetic endemism according to CANAPE scheme of Mishler et al. 2014.

cpr_classify_endem(df)

Arguments

df
Input data frame. Must have the following columns:

pe_obs_p_upper: Upper p-value comparing observed phylogenetic endemism to random values

pe_alt_obs_p_upper: Upper p-value comparing observed phylogenetic endemism on alternate tree to random values

rpe_obs_p_upper: Upper p-value comparing observed relative phylogenetic endemism to random values

Value

Dataframe with column endem_type (character) added. Values of endem_type type include paleo (paleoendemic), neo (neoendemic), not significant (what it says), mixed (mixed endemism), and super (super-endemic; both pe_obs and pe_obs_alt are highly significant).

Details

For a summary of the classification scheme, see: http://biodiverse-analysis-software.blogspot.com/2014/11/canape-categorical-analysis-of-palaeo.html

References

Mishler, B., Knerr, N., González-Orozco, C. et al. (2014) Phylogenetic measures of biodiversity and neo- and paleo-endemism in Australian Acacia. Nat Commun, 5: 4473. doi: 10.1038/ncomms5473

Examples

set.seed(12345)
data(phylocom)
rand_test <- cpr_rand_test(
  phylocom$comm, phylocom$phy,
  null_model = "curveball", metrics = c("pe", "rpe"))

#> [1] "Dropping tips from the tree because they are not present in the community data:"
#> [1] "sp16" "sp23" "sp27" "sp28" "sp30" "sp31" "sp32"

cpr_classify_endem(rand_test)

#>            pe_obs pe_rand_mean pe_rand_sd   pe_obs_z pe_obs_c_upper
#> clump1  0.1333333    0.1626069 0.02966082 -0.9869446             18
#> clump2a 0.1081761    0.1694874 0.03082893 -1.9887590              0
#> clump2b 0.1286164    0.1723491 0.03079849 -1.4199627              6
#> clump4  0.1411950    0.1702075 0.02787481 -1.0408171             13
#> even    0.2506289    0.1630692 0.02978852  2.9393787            100
#> random  0.2380503    0.1622799 0.02685895  2.8210498             99
#>         pe_obs_c_lower pe_obs_q pe_obs_p_upper pe_obs_p_lower pe_alt_obs
#> clump1              81      100           0.18           0.81  0.1472222
#> clump2a            100      100           0.00           1.00  0.1194444
#> clump2b             94      100           0.06           0.94  0.1420139
#> clump4              85      100           0.13           0.85  0.1454861
#> even                 0      100           1.00           0.00  0.2246528
#> random               1      100           0.99           0.01  0.2211806
#>         pe_alt_rand_mean pe_alt_rand_sd pe_alt_obs_z pe_alt_obs_c_upper
#> clump1         0.1624965     0.02339824   -0.6527971                 30
#> clump2a        0.1706840     0.02549820   -2.0095373                  1
#> clump2b        0.1704757     0.02347660   -1.2123480                 12
#> clump4         0.1659583     0.02137130   -0.9579306                 15
#> even           0.1653681     0.02309614    2.5668666                100
#> random         0.1650174     0.02162218    2.5974812                 99
#>         pe_alt_obs_c_lower pe_alt_obs_q pe_alt_obs_p_upper pe_alt_obs_p_lower
#> clump1                  70          100               0.30               0.70
#> clump2a                 99          100               0.01               0.99
#> clump2b                 88          100               0.12               0.88
#> clump4                  85          100               0.15               0.85
#> even                     0          100               1.00               0.00
#> random                   1          100               0.99               0.01
#>           rpe_obs rpe_rand_mean rpe_rand_sd  rpe_obs_z rpe_obs_c_upper
#> clump1  0.9056604     0.9990213  0.09849594 -0.9478653               5
#> clump2a 0.9056604     0.9920766  0.09284487 -0.9307594               6
#> clump2b 0.9056604     1.0104896  0.11063714 -0.9475050               7
#> clump4  0.9705048     1.0263369  0.11074938 -0.5041305              46
#> even    1.1156280     0.9830377  0.08516802  1.5568087              89
#> random  1.0762714     0.9830892  0.09818383  0.9490594              86
#>         rpe_obs_c_lower rpe_obs_q rpe_obs_p_upper rpe_obs_p_lower
#> clump1               88       100            0.05            0.88
#> clump2a              86       100            0.06            0.86
#> clump2b              87       100            0.07            0.87
#> clump4               54       100            0.46            0.54
#> even                 11       100            0.89            0.11
#> random               14       100            0.86            0.14
#>              endem_type
#> clump1  not significant
#> clump2a not significant
#> clump2b not significant
#> clump4  not significant
#> even              super
#> random            mixed