Classify phylogenetic endemism

Given the results of cpr_rand_test(), classifies phylogenetic endemism according to CANAPE scheme of Mishler et al. 2014.

cpr_classify_endem(df)

Arguments

df

Input data frame. Must have the following columns:

pe_obs_p_upper: Upper p-value comparing observed phylogenetic endemism to random values

pe_alt_obs_p_upper: Upper p-value comparing observed phylogenetic endemism on alternate tree to random values

rpe_obs_p_upper: Upper p-value comparing observed relative phylogenetic endemism to random values

Value

Object of class data.frame with column endem_type (character) added. Values of endem_type type include paleo (paleoendemic), neo

(neoendemic), not significant (what it says), mixed (mixed endemism), and super (super-endemic; both pe_obs and pe_obs_alt are highly significant).

Details

For a summary of the classification scheme, see: http://biodiverse-analysis-software.blogspot.com/2014/11/canape-categorical-analysis-of-palaeo.html # nolint

References

Mishler, B., Knerr, N., González-Orozco, C. et al. (2014) Phylogenetic measures of biodiversity and neo- and paleo-endemism in Australian Acacia. Nat Commun, 5: 4473. doi:10.1038/ncomms5473

Examples

set.seed(12345)
data(phylocom)
rand_test <- cpr_rand_test(
  phylocom$comm, phylocom$phy,
  null_model = "curveball", metrics = c("pe", "rpe"), n_reps = 50
)
#> Warning: Dropping tips from the tree because they are not present in the community data: 
#>  sp16, sp23, sp27, sp28, sp30, sp31, sp32
cpr_classify_endem(rand_test)
#>            pe_obs pe_rand_mean pe_rand_sd   pe_obs_z pe_obs_c_upper
#> clump1  0.1333333    0.1629245 0.02851269 -1.0378255              9
#> clump2a 0.1081761    0.1679182 0.03453727 -1.7297874              0
#> clump2b 0.1286164    0.1736981 0.03249949 -1.3871528              3
#> clump4  0.1411950    0.1679182 0.02795142 -0.9560614              6
#> even    0.2506289    0.1626918 0.02900013  3.0323008             50
#> random  0.2380503    0.1648491 0.02384301  3.0701355             50
#>         pe_obs_c_lower pe_obs_q pe_obs_p_upper pe_obs_p_lower pe_alt_obs
#> clump1              40       50           0.18           0.80  0.1472222
#> clump2a             50       50           0.00           1.00  0.1194444
#> clump2b             47       50           0.06           0.94  0.1420139
#> clump4              43       50           0.12           0.86  0.1454861
#> even                 0       50           1.00           0.00  0.2246528
#> random               0       50           1.00           0.00  0.2211806
#>         pe_alt_rand_mean pe_alt_rand_sd pe_alt_obs_z pe_alt_obs_c_upper
#> clump1         0.1641319     0.02219599   -0.7618367                 14
#> clump2a        0.1682569     0.02697502   -1.8095448                  1
#> clump2b        0.1719306     0.02484078   -1.2043369                  5
#> clump4         0.1638819     0.02212919   -0.8312928                  8
#> even           0.1652639     0.02152674    2.7588421                 50
#> random         0.1665347     0.01738758    3.1428089                 50
#>         pe_alt_obs_c_lower pe_alt_obs_q pe_alt_obs_p_upper pe_alt_obs_p_lower
#> clump1                  36           50               0.28               0.72
#> clump2a                 49           50               0.02               0.98
#> clump2b                 45           50               0.10               0.90
#> clump4                  42           50               0.16               0.84
#> even                     0           50               1.00               0.00
#> random                   0           50               1.00               0.00
#>           rpe_obs rpe_rand_mean rpe_rand_sd  rpe_obs_z rpe_obs_c_upper
#> clump1  0.9056604     0.9914623  0.10229538 -0.8387665               3
#> clump2a 0.9056604     0.9943639  0.09463083 -0.9373636               3
#> clump2b 0.9056604     1.0085663  0.10964487 -0.9385383               5
#> clump4  0.9705048     1.0265177  0.11951675 -0.4686615              24
#> even    1.1156280     0.9813165  0.08701376  1.5435668              44
#> random  1.0762714     0.9900219  0.10324640  0.8353757              43
#>         rpe_obs_c_lower rpe_obs_q rpe_obs_p_upper rpe_obs_p_lower
#> clump1               43        50            0.06            0.86
#> clump2a              45        50            0.06            0.90
#> clump2b              43        50            0.10            0.86
#> clump4               26        50            0.48            0.52
#> even                  6        50            0.88            0.12
#> random                7        50            0.86            0.14
#>              endem_type
#> clump1  not significant
#> clump2a not significant
#> clump2b not significant
#> clump4  not significant
#> even              super
#> random            super