Classify phylogenetic endemism

Given the results of cpr_rand_test(), classifies phylogenetic endemism according to CANAPE scheme of Mishler et al. 2014.

cpr_classify_endem(df)

Arguments

df
Input data frame. Must have the following columns:

pe_obs_p_upper: Upper p-value comparing observed phylogenetic endemism to random values

pe_alt_obs_p_upper: Upper p-value comparing observed phylogenetic endemism on alternate tree to random values

rpe_obs_p_upper: Upper p-value comparing observed relative phylogenetic endemism to random values

Source

Mishler, B., Knerr, N., González-Orozco, C. et al. (2014) Phylogenetic measures of biodiversity and neo- and paleo-endemism in Australian Acacia. Nat Commun, 5: 4473. doi: 10.1038/ncomms5473

Value

Dataframe with column endem_type (character) added. Values of endem_type type include paleo (paleoendemic), neo (neoendemic), not significant (what it says), mixed (mixed endemism), and super (super-endemic; both pe_obs and pe_obs_alt are highly significant).

Details

For a summary of the classification scheme, see: http://biodiverse-analysis-software.blogspot.com/2014/11/canape-categorical-analysis-of-palaeo.html

Examples

library(picante)

#> Loading required package: vegan

#> Loading required package: permute

#> Loading required package: lattice

#> This is vegan 2.5-7

#> Loading required package: nlme

data(phylocom)
rand_test <- cpr_rand_test(phylocom$sample, phylocom$phy, metrics = c("pe", "rpe"))

#> [1] "Dropping tips from the tree because they are not present in the community data:"
#> [1] "sp16" "sp23" "sp27" "sp28" "sp30" "sp31" "sp32"

cpr_classify_endem(rand_test)

#>            pe_obs pe_rand_mean pe_rand_sd   pe_obs_z pe_obs_c_upper
#> clump1  0.1333333    0.1643145 0.02859872 -1.0833049              8
#> clump2a 0.1081761    0.1655786 0.02790195 -2.0572938              1
#> clump2b 0.1286164    0.1664560 0.02986440 -1.2670478             10
#> clump4  0.1411950    0.1666478 0.02628396 -0.9683788             17
#> even    0.2506289    0.1651698 0.02847858  3.0008206            100
#> random  0.2380503    0.1718333 0.02783237  2.3791352            100
#>         pe_obs_c_lower pe_obs_q pe_obs_p_upper pe_obs_p_lower pe_alt_obs
#> clump1              92      100           0.08           0.92  0.1472222
#> clump2a             99      100           0.01           0.99  0.1194444
#> clump2b             90      100           0.10           0.90  0.1420139
#> clump4              83      100           0.17           0.83  0.1454861
#> even                 0      100           1.00           0.00  0.2246528
#> random               0      100           1.00           0.00  0.2211806
#>         pe_alt_rand_mean pe_alt_rand_sd pe_alt_obs_z pe_alt_obs_c_upper
#> clump1         0.1664306     0.02340602   -0.8206579                 17
#> clump2a        0.1676528     0.01981308   -2.4331568                  1
#> clump2b        0.1657396     0.02438609   -0.9729191                 18
#> clump4         0.1653264     0.01936701   -1.0244370                 18
#> even           0.1643542     0.02425978    2.4855384                 99
#> random         0.1704965     0.02278607    2.2243426                 99
#>         pe_alt_obs_c_lower pe_alt_obs_q pe_alt_obs_p_upper pe_alt_obs_p_lower
#> clump1                  83          100               0.17               0.83
#> clump2a                 99          100               0.01               0.99
#> clump2b                 82          100               0.18               0.82
#> clump4                  82          100               0.18               0.82
#> even                     1          100               0.99               0.01
#> random                   1          100               0.99               0.01
#>           rpe_obs rpe_rand_mean rpe_rand_sd  rpe_obs_z rpe_obs_c_upper
#> clump1  0.9056604     0.9868537  0.09668403 -0.8397806               7
#> clump2a 0.9056604     0.9854372  0.09658621 -0.8259651               6
#> clump2b 0.9056604     1.0034321  0.09869689 -0.9906266              10
#> clump4  0.9705048     1.0072990  0.09802223 -0.3753655              49
#> even    1.1156280     1.0057375  0.10475577  1.0490160              82
#> random  1.0762714     1.0080821  0.09998678  0.6819836              76
#>         rpe_obs_c_lower rpe_obs_q rpe_obs_p_upper rpe_obs_p_lower
#> clump1               85       100            0.07            0.85
#> clump2a              85       100            0.06            0.85
#> clump2b              81       100            0.10            0.81
#> clump4               51       100            0.49            0.51
#> even                 18       100            0.82            0.18
#> random               24       100            0.76            0.24
#>              endem_type
#> clump1  not significant
#> clump2a not significant
#> clump2b not significant
#> clump4  not significant
#> even              mixed
#> random            mixed