Classify statistical significance

Given the results of cpr_rand_test(), classifies statistical significance of a biodiversity metric. The null hypothesis is that observed value does not lie in the extreme of the random values.

cpr_classify_signif(df, metric, one_sided = FALSE, upper = FALSE)

Arguments

df
Input data frame.

metric
Selected metric to classify significance. May choose from pd (phylogenetic diversity), rpd (relative phylogenetic diversity), pe (phylogenentic endemism), rpe (relative phylogenetic endemism).

one_sided
Logical; is the null hypothesis one-sided? If TRUE, values will be classified as significant if they are in either the top 5% or bottom 5%. If FALSE, values will be classified as significant if they are in the top 2.5% or bottom 2.5%, combined.

upper
Logical; only applies if one_sided is TRUE. If TRUE, values in the top 5% will be classified as significant. If FALSE, values in the bottom 5% will be classified as significant.

Value

Dataframe with column added for stastistical significance of the selected metric. The new column name is the name of the metric with _signif appendend. The new column is a character that may contain the following values, depending on the null hypothesis:

< 0.01, < 0.025, > 0.975, > 0.99, not significant (two-sided)

< 0.01, < 0.05, > 0.99, > 0.95, not significant (one-sided)

Details

For metrics like pe, you probably want to consider a one-sided hypothesis testing values in the upper extreme (i.e., we are interested in areas that have higher than expected endemism). For this, you would set one_sided = TRUE, upper = TRUE. For metrics like pd, you probably want to consider a two-sided hypothesis (i.e., we are interested in areas that are either more diverse or less than diverse than expected at random). For this, set one_sided = FALSE.

Examples

library(picante)
data(phylocom)
rand_test <- cpr_rand_test(phylocom$sample, phylocom$phy, metrics = "pd")

#> [1] "Dropping tips from the tree because they are not present in the community data:"
#> [1] "sp16" "sp23" "sp27" "sp28" "sp30" "sp31" "sp32"

cpr_classify_signif(rand_test, "pd")

#>            pd_obs pd_rand_mean pd_rand_sd  pd_obs_z pd_obs_c_upper
#> clump1  0.3018868    0.4664151 0.03252358 -5.058739              0
#> clump2a 0.3207547    0.4694340 0.02891839 -5.141340              0
#> clump2b 0.3396226    0.4684906 0.03434917 -3.751704              0
#> clump4  0.4150943    0.4716981 0.03010279 -1.880350              2
#> even    0.5660377    0.4715094 0.03825590  2.470947            100
#> random  0.5094340    0.4683019 0.02880625  1.427887             83
#>         pd_obs_c_lower pd_obs_q pd_obs_p_upper pd_obs_p_lower       pd_signif
#> clump1             100      100           0.00           1.00          < 0.01
#> clump2a            100      100           0.00           1.00          < 0.01
#> clump2b            100      100           0.00           1.00          < 0.01
#> clump4              95      100           0.02           0.95 not significant
#> even                 0      100           1.00           0.00          > 0.99
#> random               3      100           0.83           0.03 not significant